% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_aae_encode.R
\name{aae_encode}
\alias{aae_encode}
\title{Adversarial Autoencoder - Encode}
\usage{
aae_encode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
a \code{autoenc_encode} object.
}
\description{
Creates an deep learning adversarial autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See example at https://nbviewer.org/github/cefet-rj-dal/daltoolbox-examples
}
