% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_aae_encode_decode.R
\name{aae_encode_decode}
\alias{aae_encode_decode}
\title{Adversarial Autoencoder - Encode}
\usage{
aae_encode_decode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
a \code{aae_encode_decode} object.
}
\description{
Creates an deep learning adversarial autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `aae_encode_decode` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/transf/aae_enc_decode.md
}
