% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_cluster.R
\name{action.cluster}
\alias{action.cluster}
\title{Action implementation for clustering}
\usage{
\method{action}{cluster}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
It simply returns NULL, which indicates that no transforms are applied
}
\description{
A default function that defines the action to proxy cluster method
}
\examples{
trans <- dal_transform()
}
