% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_transform.R
\name{action.dal_transform}
\alias{action.dal_transform}
\title{Action implementation for transform}
\usage{
\method{action}{dal_transform}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
returns a transformed data
}
\description{
A default function that defines the action to proxy transform method
}
\examples{
#See ?minmax for an example of transformation
}
