% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{action.default}
\alias{action.default}
\title{Action}
\usage{
\method{action}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
The default behavior is return the input data
}
\description{
Executes the action of an object.
}
