% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_predictor.R
\name{action.predictor}
\alias{action.predictor}
\title{Action implementation for predictor}
\usage{
\method{action}{predictor}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
It simply returns NULL, which indicates that no transforms are applied
}
\description{
A default function that defines the action to proxy predict method
}
\examples{
trans <- dal_transform()
}
