% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{action}
\alias{action}
\title{Action}
\usage{
action(obj, ...)
}
\arguments{
\item{obj}{object: a dal_base object to apply the transformation on the input dataset.}

\item{...}{optional arguments.}
}
\value{
A transformed dataset.
}
\description{
Executes the action of an object.
}
\examples{
data(iris)
result <- action(dal_base(), iris)
}
