% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_main.R
\name{adjust.factor}
\alias{adjust.factor}
\title{adjust categorical values}
\usage{
adjust.factor(value, ilevels, slevels)
}
\arguments{
\item{value}{vector to be categorized}

\item{ilevels}{\itemize{
\item order for categorical values
}}

\item{slevels}{\itemize{
\item labels for categorical values
}}
}
\value{
factor
}
\description{
adjust categorical values
}
\examples{
trans <- dal_transform()
}
