% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjust.R
\name{adjust_class_label}
\alias{adjust_class_label}
\title{adjust categorical mapping}
\usage{
adjust_class_label(x, valTrue = 1, valFalse = 0)
}
\arguments{
\item{x}{vector to be categorized}

\item{valTrue}{value to represent true}

\item{valFalse}{value to represent false}
}
\value{
an adjusted categorical mapping
}
\description{
vector \code{value} is adjusted to a categorical mapping
}
