% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_data.frame}
\alias{adjust_data.frame}
\title{Adjust input data to data frame}
\usage{
adjust_data.frame(data)
}
\arguments{
\item{data}{dataset}
}
\value{
The date argument
}
\description{
adjust_data.frame function receives
a data object and converts it to a data frame.

It returns the same input if it is already a data frame.
}
\examples{
data(iris)
df <- adjust_data.frame(iris)
}
