% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{adjust_ts_data}
\alias{adjust_ts_data}
\title{Transform the date object}
\usage{
adjust_ts_data(data)
}
\arguments{
\item{data}{dataset}
}
\value{
The date object changed
}
\description{
The first check that is done is to see if data is a matrix using the is.matrix() function. If data is not a matrix, the function converts data to a matrix using the as.matrix() function
}
