% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_main.R
\name{adjustClassLabels}
\alias{adjustClassLabels}
\title{compute categorical mapping}
\usage{
adjustClassLabels(x, valTrue = 1, valFalse = 0)
}
\arguments{
\item{x}{vector to be categorized}

\item{valTrue}{\itemize{
\item value to represent true
}}

\item{valFalse}{\itemize{
\item value to represent false
}}
}
\value{
factor
}
\description{
compute categorical mapping
}
\examples{
trans <- dal_transform()
#'@export
}
