% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_aggregation.R
\name{aggregation}
\alias{aggregation}
\title{Aggregation by groups}
\usage{
aggregation(group, ...)
}
\arguments{
\item{group}{grouping column name (string)}

\item{...}{named expressions evaluated per group}
}
\value{
returns an object of class \code{aggregation}
}
\description{
Aggregate data by a grouping attribute using named expressions.
}
\examples{
data(iris)
agg <- aggregation(
 "Species",
 mean_sepal = mean(Sepal.Length),
 sd_sepal = sd(Sepal.Length),
 n = n()
)
iris_agg <- transform(agg, iris)
iris_agg
}
