% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_adv_e.R
\name{autoenc_adv_e}
\alias{autoenc_adv_e}
\title{Adversarial Autoencoder - Encode}
\usage{
autoenc_adv_e(
  input_size,
  encoding_size,
  batch_size = 350,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
a \code{autoenc_adv_e} object.
#See an example of using \code{autoenc_adv_e} at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_adv_e.md
}
\description{
Creates an deep learning adversarial autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
