% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_base_e.R
\name{autoenc_base_e}
\alias{autoenc_base_e}
\title{Autoencoder - Encode}
\usage{
autoenc_base_e(input_size, encoding_size)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}
}
\value{
returns a \code{autoenc_base_e} object.
}
\description{
Creates a base class for autoencoder.
}
\examples{
#See an example of using `autoenc_base_e` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_base_e.md
}
