% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_denoise_ed.R
\name{autoenc_denoise_ed}
\alias{autoenc_denoise_ed}
\title{Denoising Autoencoder - Encode}
\usage{
autoenc_denoise_ed(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001,
  noise_factor = 0.3
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}

\item{noise_factor}{level of noise to be added to the data}
}
\value{
a \code{autoenc_denoise_ed} object.
}
\description{
Creates an deep learning denoising autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `autoenc_denoise_ed` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_denoise_ed.md
}
