% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_e.R
\name{autoenc_e}
\alias{autoenc_e}
\title{Autoencoder - Encode}
\usage{
autoenc_e(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{autoenc_e} object.
}
\description{
Creates an deep learning autoencoder to encode a sequence of observations.
It wraps the pytorch and reticulate libraries.
}
\examples{
#See an example of using `autoenc_e` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_e.md
}
