% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_ed.R
\name{autoenc_ed}
\alias{autoenc_ed}
\title{Autoencoder - Encode-decode}
\usage{
autoenc_ed(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{autoenc_ed} object.
}
\description{
Creates an deep learning autoencoder to encode-decode a sequence of observations.
It wraps the pytorch and reticulate libraries.
}
\examples{
#See an example of using `autoenc_ed` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_ed.md
}
