% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_autoenc_encode.R
\name{autoenc_encode}
\alias{autoenc_encode}
\title{Autoencoder - Encode}
\usage{
autoenc_encode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{autoenc_encode} object.
}
\description{
Creates an deep learning autoencoder to encode a sequence of observations.
It wraps the pytorch and reticulate libraries.
}
\examples{
#See an example of using `autoenc_encode` at this
#[link](https://github.com/cefet-rj-dal/daltoolbox/blob/main/transformation/encode.ipynb)
}
