% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_lstm_e.R
\name{autoenc_lstm_e}
\alias{autoenc_lstm_e}
\title{LSTM Autoencoder - Encode}
\usage{
autoenc_lstm_e(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 50,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{autoenc_lstm_e} object.
}
\description{
Creates an deep learning LSTM autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `autoenc_lstm_e` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_lstm_e.md
}
