% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_stacked_ed.R
\name{autoenc_stacked_ed}
\alias{autoenc_stacked_ed}
\title{Stacked Autoencoder - Encode}
\usage{
autoenc_stacked_ed(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001,
  k = 3
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}

\item{k}{number of AE layers in the stack}
}
\value{
a \code{autoenc_stacked_ed} object.
}
\description{
Creates an deep learning stacked autoencoder to encode a sequence of observations.
The autoencoder layers are based on DAL Toolbox Vanilla Autoencoder
It wraps the pytorch library.
}
\examples{
#See an example of using `autoenc_stacked_ed` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_stacked_e.md
}
