% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoenc_variational_ed.R
\name{autoenc_variational_ed}
\alias{autoenc_variational_ed}
\title{Variational Autoencoder - Encode}
\usage{
autoenc_variational_ed(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{autoenc_variational_ed} object.
}
\description{
Creates an deep learning variational autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `autoenc_variational_ed` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/autoencoder/autoenc_variational_ed.md
}
