% Generated by roxygen2: do not edit by hand
<<<<<<<< HEAD:man/c2den_encode_decode.Rd
% Please edit documentation in R/trans_c2den_encode_decode.R
\name{c2den_encode_decode}
\alias{c2den_encode_decode}
\title{Convolutional 2d Denoising Autoencoder - Encode}
\usage{
c2den_encode_decode(
========
% Please edit documentation in R/trans_cae2d_encode.R
\name{cae2d_encode}
\alias{cae2d_encode}
\title{Convolutional 2d Autoencoder - Encode}
\usage{
cae2d_encode(
>>>>>>>> conv2d_autoencoder:man/cae2d_encode.Rd
  input_size,
  encoding_size,
  mean_var_size = 6,
  batch_size = 32,
  num_epochs = 50,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
<<<<<<<< HEAD:man/c2den_encode_decode.Rd
a \code{c2den_encode_decode} object.
========
a \code{cae2d_encode_decode} object.
>>>>>>>> conv2d_autoencoder:man/cae2d_encode.Rd
}
\description{
Creates an deep learning convolutional denoising autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See example at https://nbviewer.org/github/cefet-rj-dal/daltoolbox-examples
}
