% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_cae2d_encode_decode.R
\name{cae2d_encode_decode}
\alias{cae2d_encode_decode}
\title{Convolutional 2d Autoencoder - Encode}
\usage{
cae2d_encode_decode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 50,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
a \code{cae2d_encode_decode} object.

returns a \code{cae2d_encode_decode} object.
}
\description{
Creates an deep learning convolutional autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `cae2d_encode_decode` at this
#[link](https://github.com/cefet-rj-dal/daltoolbox/blob/main/transf/cae2d_enc_decode.ipynb)
}
