% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_categ_mapping.R
\name{categ_mapping}
\alias{categ_mapping}
\title{Categorical Mapping}
\usage{
categ_mapping(attribute)
}
\arguments{
\item{attribute}{attribute target to model building.}
}
\value{
An instance of the CategoricalMapping class.
}
\description{
The CategoricalMapping class in R provides a way to map the levels of a categorical variable to new values. It is often used to recode or reclassify categorical variables in data preprocessing and data analysis tasks.
}
\examples{
trans <- dal_transform()
}
