% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_bagging.R
\name{cla_bagging}
\alias{cla_bagging}
\title{Bagging (ipred)}
\usage{
cla_bagging(attribute, nbagg = 25)
}
\arguments{
\item{attribute}{target attribute name}

\item{nbagg}{number of bootstrap aggregations}
}
\value{
returns a \code{cla_bagging} object
}
\description{
Bagging classifier using \code{ipred::bagging}.
}
\examples{
if (requireNamespace("ipred", quietly = TRUE)) {
 data(iris)
 model <- cla_bagging("Species", nbagg = 25)
 model <- fit(model, iris)
 pred <- predict(model, iris)
 table(pred, iris$Species)
}
}
