% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_boosting.R
\name{cla_boosting}
\alias{cla_boosting}
\title{Boosting (adabag)}
\usage{
cla_boosting(attribute, mfinal = 50)
}
\arguments{
\item{attribute}{target attribute name}

\item{mfinal}{number of boosting iterations}
}
\value{
returns a \code{cla_boosting} object
}
\description{
Boosting classifier using \code{adabag::boosting}.
}
\examples{
if (requireNamespace("adabag", quietly = TRUE)) {
 data(iris)
 model <- cla_boosting("Species", mfinal = 10)
 model <- fit(model, iris)
 pred <- predict(model, iris)
 table(pred, iris$Species)
}
}
