% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_dtree.R
\name{cla_dtree}
\alias{cla_dtree}
\title{Decision Tree Classification}
\usage{
cla_dtree(attribute, slevels = NULL)
}
\arguments{
\item{attribute}{The name of the attribute used as the target classification.}

\item{slevels}{The possible values for the target classification.}
}
\value{
A classification object that uses the Decision Tree algorithm for classification.
}
\description{
Creates a classification object that uses the Decision Tree algorithm for classification.
}
\details{
This function trains a Decision Tree model on the given dataset and creates a classification object that can be used to predict the target attribute values for new data points.
}
