% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_knn.R
\name{cla_knn}
\alias{cla_knn}
\title{K Nearest Neighbor Classification}
\usage{
cla_knn(attribute, slevels, k = 1)
}
\arguments{
\item{attribute}{attribute target to model building.}

\item{slevels}{Possible values for the target classification.}

\item{k}{A vector of integers indicating the number of neighbors to be considered.}
}
\value{
A classification object.
}
\description{
Classifies using the K-Nearest Neighbor algorithm.
}
\examples{
trans <- dal_transform()
}
