% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_mlp.R
\name{cla_mlp}
\alias{cla_mlp}
\title{Classification using Artificial Neural Network (ANN)}
\usage{
cla_mlp(attribute, slevels, size = NULL, decay = 0.1, maxit = 1000)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{size}{\itemize{
\item number of nodes that will be used in the hidden layer
}}

\item{decay}{\itemize{
\item how quickly it decreases in gradient descent
}}

\item{maxit}{\itemize{
\item maximun interations
}}
}
\value{
a classification object
}
\description{
Performs classification using Artificial Neural Network (ANN) algorithm
}
\examples{
trans <- dal_transform()
}
