% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_nb.R
\name{cla_nb}
\alias{cla_nb}
\title{Naive Bayes Classifier}
\usage{
cla_nb(attribute, slevels)
}
\arguments{
\item{attribute}{Name of the attribute used as the target classification.}

\item{slevels}{Possible values for the target classification.}
}
\value{
A classification object.
}
\description{
Classification using the Naive Bayes algorithm
}
\details{
This function fits a Naive Bayes model to the data provided, which can be used to classify new instances. The Naive Bayes algorithm is a probabilistic algorithm that makes predictions based on the probability of each class, given the values of the input features.
}
