% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{cla_rf}
\alias{cla_rf}
\title{Random Forest Classifier}
\usage{
cla_rf(attribute, slevels, mtry = NULL, ntree = 10)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\description{
Random Forest Classifier
}
