% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rpart.R
\name{cla_rpart}
\alias{cla_rpart}
\title{CART (rpart)}
\usage{
cla_rpart(attribute)
}
\arguments{
\item{attribute}{target attribute name}
}
\value{
returns a \code{cla_rpart} object
}
\description{
Classification tree using \code{rpart::rpart}.
}
\examples{
if (requireNamespace("rpart", quietly = TRUE)) {
 data(iris)
 model <- cla_rpart("Species")
 model <- fit(model, iris)
 pred <- predict(model, iris)
 table(pred, iris$Species)
}
}
