% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_svm.R
\name{cla_svm}
\alias{cla_svm}
\title{Support Vector Machine Classification}
\usage{
cla_svm(attribute, slevels, epsilon = 0.1, cost = 10, kernel = "radial")
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{epsilon}{\itemize{
\item parameter that controls the width of the margin around the separating hyperplane
}}

\item{cost}{\itemize{
\item parameter that controls the trade-off between having a wide margin and correctly classifying training data points
}}

\item{kernel}{\itemize{
\item the type of kernel function to be used in the SVM algorithm (linear, radial, polynomial, sigmoid)
}}
}
\value{
classification object
}
\description{
Classification using Support Vector Machine (SVM) algorithm
}
\examples{
trans <- dal_transform()
}
