% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{cla_tune}
\alias{cla_tune}
\title{Regression Tune}
\usage{
cla_tune(base_model, folds = 10, metric = "accuracy")
}
\arguments{
\item{base_model}{base model for tuning}

\item{folds}{number of folds for cross-validation}

\item{metric}{metric used to optimize}
}
\value{
a \code{cla_tune} object.
}
\description{
Regression Tune
}
\examples{
trans <- dal_transform()
}
