% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_classification.R
\name{classification}
\alias{classification}
\title{classification class}
\usage{
classification(attribute, slevels)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\value{
classification object
}
\description{
Ancestor class for classification problems
}
\examples{
trans <- dal_transform()
data(iris)
template_model <- classification("Species", levels(iris$Species))
print(template_model)
}
