% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_tune.R
\name{clu_tune}
\alias{clu_tune}
\title{Clustering Tune}
\usage{
clu_tune(base_model)
}
\arguments{
\item{base_model}{base model for tuning}
}
\value{
a \code{clu_tune} object.
}
\description{
Clustering Tune
}
\examples{
trans <- dal_transform()
}
