% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_clustering.R
\name{cluster.default}
\alias{cluster.default}
\title{cluster dataset basic method}
\usage{
\method{cluster}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments.}
}
\value{
empty data frame
}
\description{
cluster dataset basic method
}
\examples{
trans <- dal_transform()
}
