% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_main.R
\name{cluster}
\alias{cluster}
\title{cluster dataset abstract method}
\usage{
cluster(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments.}
}
\value{
obj
}
\description{
cluster dataset abstract method
}
\examples{
trans <- dal_transform()
}
