% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_main.R
\name{cluster_evaluation}
\alias{cluster_evaluation}
\title{Cluster Evaluation}
\usage{
cluster_evaluation(cluster, attribute)
}
\arguments{
\item{cluster}{A vector of integers indicating the clustering labels of the data}

\item{attribute}{A vector of integers indicating the true labels of the data}
}
\value{
A cluster_evaluation object containing the entropy metric of the clustering model
}
\description{
Evaluate the quality of a clustering model using entropy metric
}
\details{
This function computes the entropy metric of a clustering model. The entropy metric measures how well the clustering labels match the true labels. A lower entropy means that the clustering model is doing a better job at grouping similar data together.
}
