% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering Class}
\usage{
clustering()
}
\value{
clustering object
}
\description{
Ancestor class for clustering problems
}
\examples{
trans <- dal_transform()
}
