% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_dae_encode_decode.R
\name{dae_encode_decode}
\alias{dae_encode_decode}
\title{Denoising Autoencoder - Encode}
\usage{
dae_encode_decode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001,
  noise_factor = 0.3
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}

\item{noise_factor}{level of noise to be added to the data}
}
\value{
returns a \code{dae_encode_decode} object.
}
\description{
Creates an deep learning denoising autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `dae_encode_decode` at this
#[link](https://github.com/cefet-rj-dal/daltoolbox/blob/main/transf/dae_enc_decode.ipynb)
}
