% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{dal_base}
\alias{dal_base}
\title{dal_base object}
\usage{
dal_base()
}
\value{
An empty object
}
\description{
Creates a base object that will be used by other library functions
}
\examples{
trans <- dal_base()
}
