% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample.R
\name{data_sample}
\alias{data_sample}
\title{Data Sample}
\usage{
data_sample()
}
\value{
obj
}
\description{
The data_sample function in R is used to randomly sample data from a given data frame. It can be used to obtain a subset of data for further analysis or modeling.
}
\examples{
trans <- dal_transform()
}
