% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{describe.default}
\alias{describe.default}
\title{dal_base object}
\usage{
\method{describe}{default}(obj)
}
\value{
Class of the object passed as an argument in character format (string), if it is not null. If the object is null, the function returns an empty string
}
\description{
It takes as parameter the object obj. It checks if the given object is NULL and returns an empty string in that case, otherwise it returns the object's class name as a string
}
\details{
This function provides a basic description of the given object
}
