% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{describe}
\alias{describe}
\title{Describe}
\usage{
describe(obj, ...)
}
\arguments{
\item{obj}{object: the time series dataset to be described.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generates a summary of the time series dataset, including statistical
measures such as mean, variance, skewness, and kurtosis.
}
\details{
The describe function is used to generate a statistical summary of the
input time series dataset, including measures such as mean, variance, skewness,
and kurtosis. The exact measures calculated may vary depending on the specific
implementation of the describe function used.
}
