% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{describe.ts_transform}
\alias{describe.ts_transform}
\title{Describe the ts_transform function}
\usage{
\method{describe}{ts_transform}(obj)
}
\value{
The string "none"
}
\description{
The function receives as argument the variable obj
}
\details{
This function can be useful for debugging purposes or to inform the user that there is no information available about the created transform object
}
