% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_pattern_miner.R
\name{discover}
\alias{discover}
\title{Discover}
\usage{
discover(obj, ...)
}
\arguments{
\item{obj}{A \code{pattern_miner} object.}

\item{...}{Optional arguments passed to the concrete discovery method.}
}
\value{
Discovered patterns in the representation used by the mining engine.
}
\description{
Generic for pattern discovery.
}
