% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{do_fit}
\alias{do_fit}
\title{Delegate the task of adjusting the model to the specific methods of each class that implements it.}
\usage{
do_fit(obj, x, y = NULL)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}
}
\value{
fitted object
}
\description{
This function receives the obj, x and y variables as parameters
}
\examples{
trans <- dal_transform()
}
