% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_elm.R
\name{do_fit.ts_elm}
\alias{do_fit.ts_elm}
\title{Train an ELM (Extreme Learning Machine) model}
\usage{
\method{do_fit}{ts_elm}(obj, x, y)
}
\arguments{
\item{y}{}
}
\value{
The obj object updated with the trained model
}
\description{
This function takes three arguments: obj, which is an object of the ts_elm class, x, which is an array with the predictor variables, and y, which is a vector with the response variable. After training, the resulting ELM model is stored in the obj object using the model attribute
}
\details{
The function uses the implementation of the elmNNRcpp library to train the ELM model
}
