% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mlp.R
\name{do_fit.ts_mlp}
\alias{do_fit.ts_mlp}
\title{Train the ts_mlp model}
\usage{
\method{do_fit}{ts_mlp}(obj, x, y)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}
}
\value{
The ts_mlp object updated with the trained model
}
\description{
It takes three arguments: the ts_mlp object to be trained (obj), the input matrix (x) and the output vector (y)
}
