% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_rf.R
\name{do_fit.ts_rf}
\alias{do_fit.ts_rf}
\title{Fits a random forest model to time series}
\usage{
\method{do_fit}{ts_rf}(obj, x, y)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}
}
\value{
The updated ts_rf object
}
\description{
It receives as input a ts_rf object (obj), an input dataset (x) and an output dataset (y)
}
