% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_svm.R
\name{do_fit.ts_svm}
\alias{do_fit.ts_svm}
\title{Fits an SVM model to time series}
\usage{
\method{do_fit}{ts_svm}(obj, x, y)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}
}
\value{
The obj model object with the adjusted model
}
\description{
It takes as input the model object obj, the input data x and the expected outputs y
}
