% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_elm.R
\name{do_predict.ts_elm}
\alias{do_predict.ts_elm}
\title{Make predictions using a previously trained ELM model}
\usage{
\method{do_predict}{ts_elm}(obj, x)
}
\arguments{
\item{x}{}
}
\value{
A numerical object containing the predictions generated by the ELM model for the predictor variables x.
}
\description{
This function takes obj and x as arguments
}
\details{
The function first checks whether the argument x is a data.frame object and, if so, converts it to a matrix using the as.matrix function. Then the function uses the implementation of the elmNNRcpp library to make the predictions
}
