% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_rf.R
\name{do_predict.ts_rf}
\alias{do_predict.ts_rf}
\title{Make predictions on a new dataset (x) using the fitted random forest model}
\usage{
\method{do_predict}{ts_rf}(obj, x)
}
\arguments{
\item{x}{}
}
\value{
The prediction vector
}
\description{
It takes as input a ts_rf object (obj) and an input dataset (x)
}
\details{
To make the predictions, the predict() function is used, which is part of the randomForest package
}
