% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_svm.R
\name{do_predict.ts_svm}
\alias{do_predict.ts_svm}
\title{Uses an adjusted SVM model for time series to make predictions}
\usage{
\method{do_predict}{ts_svm}(obj, x)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}
}
\value{
The prediction variable
}
\description{
It takes as input the model object obj and the input data x
}
