% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{do_predict.tsreg_sw}
\alias{do_predict.tsreg_sw}
\title{Performs the prediction step of the temporal regression model}
\usage{
\method{do_predict}{tsreg_sw}(obj, x)
}
\value{
The forecast matrix
}
\description{
It receives as input the obj object that contains the trained model and the input matrix x
}
\details{
The function uses the predict() method to make predictions based on the model fitted by the do_fit() method
}
