% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_pca.R
\name{dt_pca}
\alias{dt_pca}
\title{PCA}
\usage{
dt_pca(attribute = NULL, components = NULL)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{components}{number of components for PCA}
}
\value{
obj
}
\description{
PCA (Principal Component Analysis) is a commonly used unsupervised dimensionality reduction technique in data analysis and machine learning. It is used to transform a dataset of possibly correlated variables into a new set of uncorrelated variables called principal components.
}
\examples{
trans <- dal_transform()
}
