% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_classification.R
\name{evaluate.classification}
\alias{evaluate.classification}
\title{Classification evaluation}
\usage{
\method{evaluate}{classification}(obj, data, prediction, ...)
}
\arguments{
\item{obj}{object}

\item{data}{real observations}

\item{prediction}{predicted observations}

\item{...}{optional arguments.}
}
\value{
Computed metrics
}
\description{
Evaluate major classification metrics for trained model
}
\examples{
trans <- dal_transform()
}
