% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_clustering.R
\name{evaluate.clusterer}
\alias{evaluate.clusterer}
\title{Cluster Evaluation}
\usage{
\method{evaluate}{clusterer}(obj, cluster, attribute, ...)
}
\arguments{
\item{obj}{object}

\item{cluster}{A vector of integers indicating the clustering labels of the data}

\item{attribute}{attribute target to model building}

\item{...}{optional arguments.}
}
\value{
Computed metrics
}
\description{
Evaluate the quality of a clusterer model using entropy metric
}
\examples{
trans <- dal_transform()
}
