% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_cluster.R
\name{evaluate.clustering}
\alias{evaluate.clustering}
\title{Cluster Evaluation}
\usage{
\method{evaluate}{clustering}(obj, cluster, attribute, ...)
}
\arguments{
\item{obj}{object}

\item{cluster}{A vector of integers indicating the clustering labels of the data}

\item{attribute}{attribute target to model building}

\item{...}{optional arguments.}
}
\value{
Computed metrics
}
\description{
Evaluate the quality of a clustering model using entropy metric
}
\examples{
trans <- dal_transform()
}
