% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_main.R
\name{evaluate.regression}
\alias{evaluate.regression}
\title{Regression evaluation}
\usage{
\method{evaluate}{regression}(obj, values, prediction, ...)
}
\arguments{
\item{obj}{object}

\item{values}{real observations}

\item{prediction}{predicted observations}

\item{...}{optional arguments.}
}
\value{
Computed metrics
}
\description{
Evaluate major regression metrics for trained model
}
\examples{
trans <- dal_transform()
}
