% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{evaluate.ts_reg}
\alias{evaluate.ts_reg}
\title{Calculate the Mean Squared Error (MSE) error metric and the Symmetric Mean Absolute Percentage Error (sMAPE) error metric}
\usage{
\method{evaluate}{ts_reg}(obj, values, prediction, ...)
}
\arguments{
\item{obj}{object}

\item{values}{real observations}

\item{prediction}{predicted observations}

\item{...}{optional arguments.}
}
\value{
An object that contains these metrics and their values, stored in a data frame
}
\description{
The function receives two variables as a parameter, which are values and prediction
}
