% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{evaluation.tsreg}
\alias{evaluation.tsreg}
\title{Calculate the Mean Squared Error (MSE) error metric and the Symmetric Mean Absolute Percentage Error (sMAPE) error metric}
\usage{
evaluation.tsreg(values, prediction)
}
\arguments{
\item{prediction}{}
}
\value{
An object that contains these metrics and their values, stored in a data frame
}
\description{
The function receives two variables as a parameter, which are values and prediction
}
\details{
The returned object is also assigned to the "evaluation.tsreg" class
}
