% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_feature_generation.R
\name{feature_generation}
\alias{feature_generation}
\title{Feature generation}
\usage{
feature_generation(...)
}
\arguments{
\item{...}{named expressions that compute new features}
}
\value{
returns an object of class \code{feature_generation}
}
\description{
Create new features from existing columns using named expressions.
}
\examples{
data(iris)
gen <- feature_generation(
 Sepal.Area = Sepal.Length * Sepal.Width,
 Petal.Area = Petal.Length * Petal.Width,
 Sepal.Ratio = Sepal.Length / Sepal.Width
)
iris_feat <- transform(gen, iris)
head(iris_feat)
}
