% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_feature_selection_info_gain.R
\name{feature_selection_info_gain}
\alias{feature_selection_info_gain}
\title{Feature selection by information gain}
\usage{
feature_selection_info_gain(
  attribute,
  features = NULL,
  top = NULL,
  cutoff = 0,
  bins = 3
)
}
\arguments{
\item{attribute}{target attribute name}

\item{features}{optional vector of feature names (default: all columns except \code{attribute})}

\item{top}{optional number of top features to keep}

\item{cutoff}{minimum information gain to keep a feature (default: 0)}

\item{bins}{number of quantile bins for numeric features}
}
\value{
returns an object of class \code{feature_selection_info_gain}
}
\description{
Rank and select features using information gain with optional discretization.
}
\details{
Numeric predictors are discretized by quantile bins before computing entropy-based information gain.
}
\examples{
data(iris)
fg <- feature_generation(
 IsVersicolor = ifelse(Species == "versicolor", "versicolor", "not_versicolor")
)
iris_bin <- transform(fg, iris)
iris_bin$IsVersicolor <- factor(iris_bin$IsVersicolor)
fs <- feature_selection_info_gain("IsVersicolor", top = 2)
fs <- fit(fs, iris_bin)
fs$selected
iris_fs <- transform(fs, iris_bin)
names(iris_fs)
}
