% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_dtree.R
\name{fit.cla_dtree}
\alias{fit.cla_dtree}
\title{fit dtree model}
\usage{
\method{fit}{cla_dtree}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit dtree model
}
