% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_knn.R
\name{fit.cla_knn}
\alias{fit.cla_knn}
\title{fit knn model}
\usage{
\method{fit}{cla_knn}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit knn model
}
