% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_majority.R
\name{fit.cla_majority}
\alias{fit.cla_majority}
\title{fit majority model}
\usage{
\method{fit}{cla_majority}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit majority model
}
