% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_mlp.R
\name{fit.cla_mlp}
\alias{fit.cla_mlp}
\title{fit mlp model}
\usage{
\method{fit}{cla_mlp}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit mlp model
}
