% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_nb.R
\name{fit.cla_nb}
\alias{fit.cla_nb}
\title{fit nb model}
\usage{
\method{fit}{cla_nb}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit nb model
}
