% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{fit.cla_rf}
\alias{fit.cla_rf}
\title{fit rf model}
\usage{
\method{fit}{cla_rf}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit rf model
}
