% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_svm.R
\name{fit.cla_svm}
\alias{fit.cla_svm}
\title{fit svm model}
\usage{
\method{fit}{cla_svm}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit svm model
}
