% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{fit.cla_tune}
\alias{fit.cla_tune}
\title{tune hyperparameters of ml model}
\usage{
\method{fit}{cla_tune}(obj, data, ranges, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{ranges}{hyperparameters ranges}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
tune hyperparameters of ml model for classification
}
