% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_main.R
\name{fit.classification}
\alias{fit.classification}
\title{fit classification}
\usage{
\method{fit}{classification}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
obj
}
\description{
fit abstract method for classification
}
