% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_tune.R
\name{fit.clu_tune}
\alias{fit.clu_tune}
\title{tune hyperparameters of ml model}
\usage{
\method{fit}{clu_tune}(obj, data, ranges, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{ranges}{hyperparamters ranges}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
tune hyperparameters of ml model for clusterer
}
