% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_dbscan.R
\name{fit.cluster_dbscan}
\alias{fit.cluster_dbscan}
\title{fit dbscan model}
\usage{
\method{fit}{cluster_dbscan}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
fit dbscan model
}
