% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{fit.default}
\alias{fit.default}
\title{Fit}
\usage{
\method{fit}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
The default behavior is return the object as it is
}
\description{
Fits a model from input data
}
