% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{fit.minmax}
\alias{fit.minmax}
\title{Fit MinMax}
\usage{
\method{fit}{minmax}(obj, data)
}
\description{
The fit.minmax() function in the DAL package is used to fit a min-max normalization to the input data. Min-max normalization scales the input data so that all feature values are within the range of 0 and 1. This is done by subtracting the minimum value of each feature from all values of that feature and then dividing by the range of that feature. The resulting normalization parameters are saved in the input object, and the modified object is returned.
}
\details{
The fit.minmax() function takes two arguments: obj, which is an object of class minmax returned by the minmax() function, and data, which is a data frame containing the input data to be normalized. The function first creates a data frame minmax with the same number of columns as data and sets the values of the first row to 1 for columns that contain numeric data and 0 otherwise. The second and third rows of minmax are set to NA initially. The column and row names of minmax are set to the column names and "numeric", "max", and "min", respectively.
}
