% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_outliers.R
\name{fit.outliers}
\alias{fit.outliers}
\title{Fit outliers object}
\usage{
\method{fit}{outliers}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
An updated 'outliers' object, containing the lower and upper bounds for each numerical variable
}
\description{
Calculates upper and lower bounds based on alpha value
}
\examples{
trans <- dal_transform()
}
