% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_prediction.R
\name{fit.prediction}
\alias{fit.prediction}
\title{fit predict}
\usage{
\method{fit}{prediction}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
obj
}
\description{
fit abstract method for predict
}
