% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tune.R
\name{fit.reg_tune}
\alias{fit.reg_tune}
\title{tune hyperparameters of ml model (regression)}
\usage{
\method{fit}{reg_tune}(obj, data, ranges, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{ranges}{hyperparamters ranges}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
tune hyperparameters of ml model for regression
}
