% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_an.R
\name{fit.ts_an}
\alias{fit.ts_an}
\title{Fits the "ts_an" normalization model to a training dataset}
\usage{
\method{fit}{ts_an}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
The adjusted "ts_an" object
}
\description{
It takes two arguments: obj and data. It computes the moving average of the input data using the "ma.ts_an" function, subtracts the moving average from the input data, and scales the data so that the values are in the range between 0 and 1
}
\examples{
trans <- dal_transform()
}
