% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{fit.ts_diff}
\alias{fit.ts_diff}
\title{Fit the object created by the ts_diff function to the training data}
\usage{
\method{fit}{ts_diff}(obj, data)
}
\value{
The updated obj object
}
\description{
It takes two arguments: obj and data
}
\details{
The fit.ts_diff function performs the transformation on the data data matrix that represents the difference between each pair of consecutive observations of each variable. The function then calls the fit.ts_gminmax function to normalize the transformed data matrix using the Min-Max normalization technique
}
