% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_gminmax.R
\name{fit.ts_gminmax}
\alias{fit.ts_gminmax}
\title{Fit global min max parameters for normalization}
\usage{
\method{fit}{ts_gminmax}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
The updated obj object
}
\description{
Fit global min max parameters for normalization
}
\examples{
trans <- dal_transform()
}
