% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_regsw.R
\name{fit.ts_regsw}
\alias{fit.ts_regsw}
\title{Performs the adjustment (training) of a regression model in time series}
\usage{
\method{fit}{ts_regsw}(obj, x, y, ...)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}

\item{...}{optional arguments}
}
\value{
The updated obj object
}
\description{
It takes as input an obj object of the ts_regsw class, which contains the model settings as well as the x and y data to fit the model
}
\examples{
trans <- dal_transform()
}
