% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{fit.ts_swminmax}
\alias{fit.ts_swminmax}
\title{adjust the obj object parameters for the "sw-min-max" normalization method}
\usage{
\method{fit}{ts_swminmax}(obj, data)
}
\arguments{
\item{data}{}
}
\value{
The updated obj object
}
\description{
It takes as parameters obj object and the data
}
\details{
The function checks if the option to remove outliers was selected (obj$remove_outliers) and, if it is true, adjusts the outliers through the outliers() function and then applies the transformation to data through the transform() function
}
