% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{fit.zscore}
\alias{fit.zscore}
\title{Fit Z-score}
\usage{
\method{fit}{zscore}(obj, data, ...)
}
\arguments{
\item{obj}{zscore object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted zscore object
}
\description{
Compute normalization parameters
}
\examples{
trans <- dal_transform()
}
