% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_fit.R
\name{fit_curvature}
\alias{fit_curvature}
\title{Fit}
\usage{
fit_curvature()
}
\description{
In R, the fit function is a generic function used to fit a model to a data set. It is a fundamental function in many machine learning algorithms and statistical models. The fit function takes a data set and a model specification as input and returns an object containing the fitted model parameters and other relevant information
}
\details{
The exact arguments and behavior of the fit function will depend on the specific model being fit. However, the fit function typically takes a data set as its first argument, and one or more additional arguments specifying the model to be fit. The function then performs an optimization algorithm to find the model parameters that best fit the data.
}
