% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_fit.R
\name{fit_curvature_max}
\alias{fit_curvature_max}
\title{fit_curvature_max}
\usage{
fit_curvature_max()
}
\value{
Returns an object of class fit_curvature_max, which inherits from the fit_curvature and dal_transform classes.
The object contains a list with the following elements:
\itemize{
\item func: The function used to calculate the curvature, which is set to max.
\item curvature: The maximum curvature value calculated during the fitting process.
}
}
\description{
This function creates an object for fitting a curvature function that uses the maximum curvature value.
The resulting object can be used to transform a data set using the max curvature value.
}
\details{
The fit_curvature_max function is an extension of the fit_curvature function, which creates an object for fitting a curvature function based on a specified input function.
The fit_curvature_max function sets the input function as max, resulting in a curvature function that calculates the maximum curvature value.
The resulting object can be used to transform a data set by scaling the data according to the maximum curvature value, using the transform method.
}
\examples{
trans <- dal_transform()
}
