% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_fit.R
\name{fit_curvature_min}
\alias{fit_curvature_min}
\title{fit_curvature_min}
\usage{
fit_curvature_min()
}
\value{
A list object with the optimal cut point, the values of the curvature function, the input dataset, and the function used to select the optimal cut point (minimum value).
}
\description{
The function fit_curvature_min() fits a curvature function to a dataset, and returns a list object with the results. This function is similar to fit_curvature(), but it selects the minimum value of the curvature function as the optimal cut point.
}
\details{
The fit_curvature_min() function uses the fit_curvature() function to fit a curvature function to the dataset. It then selects the minimum value of the curvature function as the optimal cut point, and returns a list object with the following elements:
\itemize{
\item cutoff: the optimal cut point.
\item curvature: the values of the curvature function for each possible cut point.
\item data: the input dataset.
\item func: the function used to select the optimal cut point (minimum value).
}
}
\examples{
trans <- dal_transform()

}
