% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_norm_fixed_zscore.R
\name{fixed_zscore}
\alias{fixed_zscore}
\title{Fixed Z-score normalization}
\usage{
fixed_zscore()
}
\value{
returns the z-score transformation object
}
\description{
Integrated scale data using z-score and minmax normalization.

\eqn{zscore = (x - mean(x))/sd(x)}
}
\examples{
data(iris)
head(iris)

trans <- fixed_zscore()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
head(tiris)

itiris <- inverse_transform(trans, tiris)
head(itiris)
}
