% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{inverse_transform.default}
\alias{inverse_transform.default}
\title{dal_base object}
\usage{
\method{inverse_transform}{default}(obj, ...)
}
\arguments{
\item{obj, }{...}
}
\value{
Simply returns NULL
}
\description{
It receives as parameter the object obj, ...
}
\details{
It is used when no specific method is defined for a specific dal_transform object and the generic function inverse_transform() is called
}
