% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_minmax.R
\name{inverse_transform.minmax}
\alias{inverse_transform.minmax}
\title{inverse transform min-max normalization}
\usage{
\method{inverse_transform}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{min-max transformation object}

\item{data}{transformed dataset}

\item{...}{optional arguments}
}
\value{
inverse transformed dataset
}
\description{
The inverse transform of normalized data to original space
}
\examples{
# see ?minmax
}
