% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Inverse Transform}
\usage{
inverse_transform(obj, ...)
}
\arguments{
\item{obj}{object: The transformed time series dataset.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The time series dataset in its original scale.
}
\description{
Reverses the transformation applied to a time series dataset using the transform() function.
}
\details{
This function is used to convert the transformed data back to its original scale, which can be useful for evaluating the performance of a time series model or visualizing the original data alongside the transformed data.
}
