% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{inverse_transform.ts_an}
\alias{inverse_transform.ts_an}
\title{Perform the inversion of the transformation performed by the transform.t_an function}
\usage{
\method{inverse_transform}{ts_an}(obj, data, x = NULL)
}
\arguments{
\item{x}{}
}
\value{
The final result of the inverse transformation
}
\description{
It takes 3 arguments: obj, data and x
}
\details{
If a vector x is given as input, the function performs the inverse transformation on that vector. Otherwise, if a data matrix is given as input, the function performs the inverse transformation on each column of that matrix
}
