% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{inverse_transform.ts_diff}
\alias{inverse_transform.ts_diff}
\title{Revert the transform applied in the transform.ts_diff function}
\usage{
\method{inverse_transform}{ts_diff}(obj, data, x = NULL)
}
\value{
If the argument "x" is provided, the function returns the value corresponding to "x" in the original scale. Otherwise, the function returns the complete "data" time series at the original scale
}
\description{
It takes as parameters the variables obj, data and x
}
\details{
If the argument x is given, the function will just apply the inverse transform to that single value. Otherwise, it will apply the inverse transform to all data in data
}
