% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{inverse_transform.ts_gminmax}
\alias{inverse_transform.ts_gminmax}
\title{Inverse of function "transform.ts_gminmax}
\usage{
\method{inverse_transform}{ts_gminmax}(obj, data, x = NULL)
}
\arguments{
\item{x}{}
}
\value{
The function returns the corresponding array of original values; if the input is a single normalized value, the function returns the corresponding non-normalized value
}
\description{
It receives as input an "obj" object that contains the minimum value (gmin) and the maximum value (gmax) of the input data that were normalized with the "transform.ts_gminmax" function
}
\details{
It applies the inverse operation to return the original (non-normalized) values of the input data, i.e. if the input is a vector of normalized data
}
