% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{inverse_transform.ts_swminmax}
\alias{inverse_transform.ts_swminmax}
\title{Do the inverse transformation of normalized data}
\usage{
\method{inverse_transform}{ts_swminmax}(obj, data, x = NULL)
}
\value{
If the parameter x is provided, the function returns the denormalized value of x. Otherwise, the function assumes that data is an array of normalized data and returns the corresponding denormalized array
}
\description{
It takes as parameters the variables obj, data, x
}
\details{
The function uses the technique of "min-max scaling" per line (or sliding window)
}
