% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{inverse_transform.zscore}
\alias{inverse_transform.zscore}
\title{Inverse Transform z-score}
\usage{
\method{inverse_transform}{zscore}(obj, data, ...)
}
\arguments{
\item{obj}{z-score object}

\item{data}{transformed dataset}

\item{...}{optional arguments}
}
\value{
reverse transformed dataset
}
\description{
The inverse transform of normalized data to original space
}
\examples{
trans <- dal_transform()
}
