% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample.R
\name{k_fold}
\alias{k_fold}
\title{k-fold sampling}
\usage{
k_fold(obj, data, k)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{k}{number of folds}
}
\value{
k folds
}
\description{
k-fold sampling
}
\examples{
trans <- dal_transform()
}
