% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lae_encode_decode.R
\name{lae_encode_decode}
\alias{lae_encode_decode}
\title{LSTM Autoencoder - Decode}
\usage{
lae_encode_decode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 50,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{lae_encode_decode} object.
}
\description{
Creates an deep learning LSTM autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `lae_encode_decode` at this
#[link](https://github.com/cefet-rj-dal/daltoolbox/blob/main/transf/lae_enc_decode.ipynb)
}
