% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ma.ts_an}
\alias{ma.ts_an}
\title{Removes the first few columns from a data array}
\usage{
ma.ts_an(obj, data, func)
}
\arguments{
\item{func}{}
}
\description{
It takes as parameters the variables obj, data, func
}
\details{
The function func is applied to the data matrix along the rows (axis 1) using the function apply(), with the option na.rm = TRUE to remove missing values
}
