% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{MSE.ts}
\alias{MSE.ts}
\title{MSE}
\usage{
MSE.ts(actual, prediction)
}
\arguments{
\item{actual}{real observations}

\item{prediction}{predicted observations}
}
\value{
returns a number, which is the calculated MSE
}
\description{
Compute the mean squared error (MSE) between actual values and forecasts of a time series
}
