% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{MSE.tsreg}
\alias{MSE.tsreg}
\title{Calculate the mean squared error (MSE) between actual values and forecasts of a time series}
\usage{
MSE.tsreg(actual, prediction)
}
\arguments{
\item{prediction}{}
}
\value{
A number, which is the calculated MSE
}
\description{
The function receives two variables as a parameter, which are actual and prediction
}
\details{
MSE is a common measure of performance for time series forecasting models, which represents the mean of the squared differences between the actual values and the forecasts
}
