% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_na_removal.R
\name{na_removal}
\alias{na_removal}
\title{Missing value removal}
\usage{
na_removal()
}
\value{
returns an object of class \code{na_removal}
}
\description{
Remove rows (or elements) that contain missing values.
}
\details{
For data frames or matrices, removes rows with any NA. For vectors, removes NA values.
}
\examples{
data(iris)
iris.na <- iris
iris.na$Sepal.Length[2] <- NA
obj <- na_removal()
iris.clean <- transform(obj, iris.na)
nrow(iris.clean)
}
