% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{optimize.default}
\alias{optimize.default}
\title{dal_base object}
\usage{
\method{optimize}{default}(obj)
}
\value{
Simply returns the object
}
\description{
It receives as parameter the obj object
}
\details{
Generic method that calls a specific implementation to optimize an object, depending on its class. When there is no specific implementation defined for the object's class, the generic method optimize.default is called
}
