% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{optimize}
\alias{optimize}
\title{Allow different types of objects to be optimized using different optimization algorithms}
\usage{
optimize(obj, ...)
}
\arguments{
\item{obj}{object: .}

\item{...}{further arguments passed to or from other methods.}
}
\description{
It uses the "UseMethod" method to call a specific optimization function for the object type that is passed to "obj"
}
\details{
This function is a generic wrapper for calling specific optimization functions for a given type of object in R.
}
