% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_outliers.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(alpha = 1.5)
}
\arguments{
\item{alpha}{}
}
\value{
An outlier object
}
\description{
This R function defines an object of class "outliers" that can be used for outlier detection in a dataset.
}
\details{
The outliers function has an optional parameter alpha that sets the threshold for identifying outliers.
The default value for alpha is 1.5, which is a common value used in many outlier detection methods.
The function creates an object using the dal_transform function, which must be defined elsewhere in the code or in an R package.
The object created by the dal_transform function can be a base class object "dal_transform",
which may include additional properties and methods for data transformation.
}
