% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.boxplot.class}
\alias{plot.boxplot.class}
\title{plot boxplot per class}
\usage{
\method{plot}{boxplot.class}(data, class_label, label_x = "", label_y = "", colors = NULL)
}
\value{
plot
}
\description{
plot boxplot per class
}
