% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.density.class}
\alias{plot.density.class}
\title{plot scatter}
\usage{
plot.density.class(
  data,
  class_label,
  label_x = "",
  label_y = "",
  colors = NULL,
  bin = NULL,
  alpha = 0.5
)
}
\value{
plot
}
\description{
plot scatter
}
