% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.density}
\alias{plot.density}
\title{plot density}
\usage{
\method{plot}{density}(data, label_x = "", label_y = "", colors = NULL, bin = NULL, alpha = 0.25)
}
\value{
plot
}
\description{
plot density
}
