% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.dotchar}
\alias{plot.dotchar}
\title{plot scatter}
\usage{
plot.dotchar(
  data,
  colors,
  colorline = "lightgray",
  xlabel = "",
  ylabel = "",
  sorting = "ascending"
)
}
\value{
plot
}
\description{
plot scatter
}
