% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_fit.R
\name{plot.fit_curvature}
\alias{plot.fit_curvature}
\title{plot.fit_curvature}
\usage{
plot.fit_curvature(obj, y, res)
}
\arguments{
\item{obj}{A fit_curvature object}

\item{y}{A numeric vector of data points to be plotted}

\item{res}{A list containing the results of fitting the data using the fit_curvature object}
}
\value{
None
}
\description{
A function to plot a fit_curvature object
}
\details{
This function takes a fit_curvature object as input and plots the corresponding data.
The points where the curvature is maximum or minimum are highlighted in red.
}
