% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.series_second_axis}
\alias{plot.series_second_axis}
\title{plot scatter}
\usage{
plot.series_second_axis(
  data,
  label_x = "",
  label_y = "",
  label_z = "",
  colors = c("blue", "red")
)
}
\value{
plot
}
\description{
plot scatter
}
