% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_boxplot_class}
\alias{plot_boxplot_class}
\title{plot boxplot per class}
\usage{
plot_boxplot_class(
  data,
  class_label,
  label_x = "",
  label_y = "",
  colors = NULL
)
}
\value{
plot
}
\description{
plot boxplot per class
}
