% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_density.class}
\alias{plot_density.class}
\title{plot density per class}
\usage{
plot_density.class(
  data,
  class_label,
  label_x = "",
  label_y = "",
  colors = NULL,
  bin = NULL,
  alpha = 0.5
)
}
\value{
plot
}
\description{
plot density per class
}
