% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_density}
\alias{plot_density}
\title{plot density}
\usage{
plot_density(
  data,
  label_x = "",
  label_y = "",
  colors = NULL,
  bin = NULL,
  alpha = 0.25
)
}
\value{
plot
}
\description{
plot density
}
