% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_groupedbar}
\alias{plot_groupedbar}
\title{plot grouped bar}
\usage{
plot_groupedbar(data, label_x = "", label_y = "", colors = NULL, alpha = 1)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{alpha}{level of transparency}
}
\value{
ggplot graphic
}
\description{
plot grouped bar
}
\examples{
trans <- dal_transform()
}
