% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_hist}
\alias{plot_hist}
\title{plot histogram}
\usage{
plot_hist(data, label_x = "", label_y = "", color = "white", alpha = 0.25)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{color}{color vector}

\item{alpha}{transparency level}
}
\value{
ggplot graphic
}
\description{
plot histogram
}
\examples{
trans <- dal_transform()
}
