% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{plot lollipop}
\usage{
plot_lollipop(
  data,
  colors,
  xlabel = "",
  ylabel = "",
  size_text = 3,
  size_ball = 8,
  alpha_ball = 0.2,
  min_value = 0,
  max_value_gap = 1,
  flip = TRUE
)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{colors}{color vector}

\item{xlabel}{x-axis label}

\item{ylabel}{y-axis label}

\item{size_text}{size of text}

\item{size_ball}{size of ball}

\item{alpha_ball}{transparency of ball}

\item{min_value}{minimum value}

\item{max_value_gap}{maximum value gap}

\item{flip}{flip axis}
}
\value{
ggplot graphic
}
\description{
plot lollipop
}
\examples{
trans <- dal_transform()
}
