% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plot_pair.R
\name{plot_pair}
\alias{plot_pair}
\title{Plot scatter matrix}
\usage{
plot_pair(data, cnames, title = NULL, clabel = NULL, colors = NULL)
}
\arguments{
\item{data}{data.frame}

\item{cnames}{column names to include}

\item{title}{optional title}

\item{clabel}{optional class label column name}

\item{colors}{optional vector of colors for classes}
}
\value{
returns a ggplot2::ggplot graphic
}
\description{
Scatter matrix using GGally::ggpairs with optional class coloring.
}
\examples{
data(iris)
grf <- plot_pair(iris, cnames = colnames(iris)[1:4], title = "Iris")
print(grf)
}
