% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plot_parallel.R
\name{plot_parallel}
\alias{plot_parallel}
\title{Plot parallel coordinates}
\usage{
plot_parallel(data, columns, group, colors = NULL, title = NULL)
}
\arguments{
\item{data}{data.frame}

\item{columns}{numeric columns to include (indices or names)}

\item{group}{grouping column (index or name)}

\item{colors}{optional vector of colors for groups}

\item{title}{optional title}
}
\value{
returns a ggplot2::ggplot graphic
}
\description{
Parallel coordinates plot using GGally::ggparcoord.
}
\examples{
data(iris)
grf <- plot_parallel(iris, columns = 1:4, group = 5)
plot(grf)
}
