% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pieplot}
\alias{plot_pieplot}
\title{plot pie}
\usage{
plot_pieplot(
  data,
  label_x = "",
  label_y = "",
  colors = NULL,
  textcolor = "white",
  bordercolor = "black"
)
}
\value{
plot
}
\description{
plot pie
}
