% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_points}
\alias{plot_points}
\title{Plot points}
\usage{
plot_points(data, label_x = "", label_y = "", colors = NULL)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
returns a ggplot graphic
}
\description{
This function creates a scatter plot using ggplot2.
}
\examples{
x <- seq(0, 10, 0.25)
data <- data.frame(x, sin=sin(x), cosine=cos(x)+5)
head(data)

grf <- plot_points(data, colors=c("red", "green"))
plot(grf)
}
