% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_series}
\alias{plot_series}
\title{plot series}
\usage{
plot_series(data, label_x = "", label_y = "", colors = NULL)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
plot
}
\description{
plot series
}
\examples{
trans <- dal_transform()
}
