% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ts_pred}
\alias{plot_ts_pred}
\title{Plot a time series chart}
\usage{
plot_ts_pred(
  x = NULL,
  y,
  yadj,
  ypred = NULL,
  label_x = "",
  label_y = "",
  color = "black",
  color_adjust = "blue",
  color_prediction = "green"
)
}
\arguments{
\item{x}{time index}

\item{y}{time series}

\item{yadj}{adjustment of time series}

\item{ypred}{prediction of the time series}

\item{label_x}{x-axis title}

\item{label_y}{y-axis title}

\item{color}{color for the time series}

\item{color_adjust}{color for the adjusted values}

\item{color_prediction}{color for the predictions}
}
\description{
The function receives six variables as a parameter, which are obj and y, yadj, main and xlabels. The graph is plotted with 3 lines: the original series (in black), the adjusted series (in blue) and the predicted series (in green)
}
