% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_majority.R
\name{predict.cla_majority}
\alias{predict.cla_majority}
\title{predict data from input}
\usage{
\method{predict}{cla_majority}(object, x, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
predicted values
}
\description{
predict data from input
}
