% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{predict.cla_rf}
\alias{predict.cla_rf}
\title{predict data from input}
\usage{
\method{predict}{cla_rf}(object, x, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
predicted values
}
\description{
predict data from input
}
