% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_knn.R
\name{predict.reg_knn}
\alias{predict.reg_knn}
\title{predict data from input}
\usage{
\method{predict}{reg_knn}(object, x, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
predicted values
}
\description{
predict data from input
}
