% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{predict.ts_reg}
\alias{predict.ts_reg}
\title{predict data from input}
\usage{
\method{predict}{ts_reg}(object, x, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
predicted values
}
\description{
predict data from input
}
\examples{
trans <- dal_transform()
}
