% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_regsw.R
\name{predict.ts_regsw}
\alias{predict.ts_regsw}
\title{Performs prediction of values for a time series}
\usage{
\method{predict}{ts_regsw}(object, x, steps_ahead = 1, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{steps_ahead}{number of step ahead for prediction}

\item{...}{optional arguments}
}
\value{
A vector with forecasts for each time period
}
\description{
The function receives three variables as a parameter, which are obj, x and steps_ahead
}
\examples{
trans <- dal_transform()
}
