% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{predict.tsreg}
\alias{predict.tsreg}
\title{Returns the last observed value}
\usage{
\method{predict}{tsreg}(obj, x)
}
\arguments{
\item{x}{}
}
\value{
The last column of the matrix x, that is, the value corresponding to the last time period in the time series
}
\description{
Takes two arguments: the object obj and an array x
}
\details{
This function does not do any forecasting or statistical modeling of the time series, it simply returns the last observed value
}
