% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_arima.R
\name{predict.tsreg_arima}
\alias{predict.tsreg_arima}
\title{predict data from input}
\usage{
\method{predict}{tsreg_arima}(object, x, y = NULL, steps_ahead = NULL, ...)
}
\arguments{
\item{object}{object}

\item{x}{input variable}

\item{y}{y modeling data}

\item{steps_ahead}{step ahead prediction}

\item{...}{optional arguments}
}
\value{
predicted values
}
\description{
predict data from input
}
