% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_prediction.R
\name{prediction}
\alias{prediction}
\title{DAL Predict}
\usage{
prediction()
}
\value{
a prediction object
}
\description{
Ancestor class for regression and classification
}
\examples{
trans <- prediction()
}
