% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_predictor.R
\name{predictor}
\alias{predictor}
\title{DAL Predict}
\usage{
predictor()
}
\value{
a predictor object
}
\description{
Ancestor class for regression and classification
It provides basis for fit and predict methods.
Besides, action method proxies to predict.

An example of learner is a decision tree (cla_dtree)
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_dtree("Species", slevels)
model <- fit(model, iris)
prediction <- predict(model, iris)
predictand <- adjust_class_label(iris[,"Species"])
train_eval <- evaluate(model, predictand, prediction)
train_eval$metrics
}
