% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{R2.ts}
\alias{R2.ts}
\title{R2}
\usage{
R2.ts(actual, prediction)
}
\arguments{
\item{actual}{real observations}

\item{prediction}{predicted observations}
}
\value{
returns a number, which is the calculated R2
}
\description{
Compute the R-squared (R2) between actual values and forecasts of a time series
}
