% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_dtree.R
\name{reg_dtree}
\alias{reg_dtree}
\title{Decision Tree for Regression}
\usage{
reg_dtree(attribute)
}
\arguments{
\item{attribute}{attribute target to model building}
}
\value{
obj
}
\description{
Decision Tree for Regression
}
\examples{
trans <- dal_transform()
}
