% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_knn.R
\name{reg_knn}
\alias{reg_knn}
\title{K-Nearest Neighbors (KNN) Regression}
\usage{
reg_knn(attribute, k)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{k}{number of k neighbors}
}
\value{
obj
}
\description{
K-Nearest Neighbors (KNN) Regression
}
\examples{
trans <- dal_transform()
}
