% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rf.R
\name{reg_rf}
\alias{reg_rf}
\title{Random Forest Regression}
\usage{
reg_rf(attribute, mtry = NULL, ntree = 10)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{mtry}{number of attributes to build trees}

\item{ntree}{number of trees}
}
\value{
obj
}
\description{
Random Forest Regression
}
\examples{
trans <- dal_transform()
}
