% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_svm.R
\name{reg_svm}
\alias{reg_svm}
\title{Support Vector Machine (SVM) Regression}
\usage{
reg_svm(attribute, epsilon = 0.1, cost = 10, kernel = "radial")
}
\arguments{
\item{attribute}{attribute target to model building}

\item{epsilon}{error threshold}

\item{cost}{cost}

\item{kernel}{SVM kernel (linear, radial, polynomial, sigmoid)}
}
\value{
obj
}
\description{
Support Vector Machine (SVM) Regression
}
\examples{
trans <- dal_transform()
}
