% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tune.R
\name{reg_tune}
\alias{reg_tune}
\title{Regression Tune}
\usage{
reg_tune(base_model, folds = 10)
}
\arguments{
\item{base_model}{base model for tuning}

\item{folds}{number of folds for cross-validation}
}
\value{
a \code{reg_tune} object.
}
\description{
Regression Tune
}
\examples{
# preparing dataset for random sampling
data(Boston)
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, Boston)
train <- sr$train
test <- sr$test

# hyper parameter setup
tune <- reg_tune(reg_svm("medv"))
ranges <- list(epsilon=seq(0,1,0.25), cost=seq(25,100,25), kernel = c("radial"))

# hyper parameter optimization
model <- fit(tune, train, ranges)

test_prediction <- predict(model, test)
test_predictand <- test[,"medv"]
test_eval <- evaluate(model, test_predictand, test_prediction)
test_eval$metrics
}
