% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{register_log.default}
\alias{register_log.default}
\title{dal_base object}
\usage{
\method{register_log}{default}(obj, msg = "")
}
\arguments{
\item{obj}{object: .}
}
\value{
The obj object updated with log record information
}
\description{
It takes as parameters the obj object and the variable msg. This function logs a log message to an object and returns the updated object itself
}
\details{
If the message is not provided, the function uses the describe function to get a description of the object's class as a log message. The function then calculates the time since the last time logging was started using the difftime function and stores the time in obj$log_time
}
