% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_sample_groups.R
\name{sample_groups}
\alias{sample_groups}
\title{Group sampling}
\usage{
sample_groups(attribute, n_groups)
}
\arguments{
\item{attribute}{group-defining attribute name}

\item{n_groups}{number of groups to sample}
}
\value{
returns an object of class \code{sample_groups}
}
\description{
Sample entire groups defined by a categorical attribute.
In sampling theory, this design is known as cluster sampling
(also called one-stage cluster sampling or sampling by groups).
The groups are assumed to be pre-defined in the data; this function
does not infer groups with clustering algorithms such as k-means.
}
\examples{
data(iris)
sc <- sample_groups("Species", n_groups = 2)
iris_sc <- transform(sc, iris)
table(iris_sc$Species)
}
