% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{select_hyper.cla_tune}
\alias{select_hyper.cla_tune}
\title{selection of hyperparameters}
\usage{
\method{select_hyper}{cla_tune}(obj, hyperparameters)
}
\arguments{
\item{obj}{object}

\item{hyperparameters}{hyperparameters dataset}
}
\value{
optimized key number of hyperparameters
}
\description{
selection of hyperparameters (maximizing classification metric)
}
