% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{select_hyper}
\alias{select_hyper}
\title{select hyper parameters}
\usage{
select_hyper(obj, hyperparameters)
}
\arguments{
\item{obj}{object: .}

\item{hyperparameters}{data set with hyper parameters.}
}
\description{
select hyper parameters
}
