% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tune.R
\name{select_hyper.reg_tune}
\alias{select_hyper.reg_tune}
\title{selection of hyperparameters}
\usage{
\method{select_hyper}{reg_tune}(obj, hyperparameters)
}
\arguments{
\item{obj}{object}

\item{hyperparameters}{hyperparameters dataset}
}
\value{
optimized key number of hyperparameters
}
\description{
selection of hyperparameters (minimizing regression error)
}
