% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_tune.R
\name{select_hyper.ts_tune}
\alias{select_hyper.ts_tune}
\title{selection of hyperparameters (time series)}
\usage{
\method{select_hyper}{ts_tune}(obj, hyperparameters)
}
\arguments{
\item{obj}{object}

\item{hyperparameters}{hyperparameters dataset}
}
\value{
optimized key number of hyperparameters
}
\description{
selection of hyperparameters (minimizing error)
}
