% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_knn.R
\name{set_params.cla_knn}
\alias{set_params.cla_knn}
\title{Set parameters values for cla_knn}
\usage{
\method{set_params}{cla_knn}(obj, params)
}
\value{
The cla_knn object updated with the new parameter values
}
\description{
It receives as input a ts_rf object (obj) and a set of parameters (params)
}
\details{
If the parameter set contains an entry for nodesize, the corresponding value is assigned to the ts_rf object. Likewise, if the parameter set contains an entry for ntree, the corresponding value is assigned to the ts_rf object
}
