% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_mlp.R
\name{set_params.cla_mlp}
\alias{set_params.cla_mlp}
\title{Set parameters values for cla_mlp}
\usage{
\method{set_params}{cla_mlp}(obj, params)
}
\value{
The cla_mlp object updated with the new parameter values
}
\description{
It receives as input a cla_mlp object (obj) and a set of parameters (params)
}
