% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{set_params.cla_rf}
\alias{set_params.cla_rf}
\title{Set parameters values for cla_rf}
\usage{
\method{set_params}{cla_rf}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
The cla_rf object updated with the new parameter values
}
\description{
It receives as input a cla_rf object (obj) and a set of parameters (params)
}
