% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_svm.R
\name{set_params.cla_svm}
\alias{set_params.cla_svm}
\title{Set parameters values for reg_svm}
\usage{
\method{set_params}{cla_svm}(obj, params)
}
\value{
The reg_svm object updated with the new parameter values
}
\description{
It receives as input a reg_svm object (obj) and a set of parameters (params)
}
