% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_kmeans.R
\name{set_params.cluster_kmeans}
\alias{set_params.cluster_kmeans}
\title{Updates the parameters of cluster_kmeans}
\usage{
\method{set_params}{cluster_kmeans}(obj, params)
}
\arguments{
\item{obj}{cluster_kmeans object}

\item{params}{parameters}
}
\value{
cluster_kmeans object
}
\description{
It takes as input the obj model object and a set of params parameters to update.
}
