% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{set_params.dal_base}
\alias{set_params.dal_base}
\title{Assign parameters}
\usage{
\method{set_params}{dal_base}(obj, params)
}
\arguments{
\item{obj}{object of class dal_base}

\item{params}{parameters to set obj}
}
\value{
obj with parameters set
}
\description{
set_params function assigns all parameters to the attributes presented in the object.
It returns the object with the parameters set.
}
