% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{set_params.default}
\alias{set_params.default}
\title{Defines a default method for the set_params function}
\usage{
\method{set_params}{default}(obj, params)
}
\arguments{
\item{params}{}
}
\value{
The obj object
}
\description{
Defines a default method for the set_params function
}
