% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{set_params}
\alias{set_params}
\title{Update an object's parameters}
\usage{
set_params(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\description{
The function receives the obj and params variables as parameters
}
\examples{
trans <- dal_transform()
}
