% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_knn.R
\name{set_params.reg_knn}
\alias{set_params.reg_knn}
\title{Set parameters values for reg_knn}
\usage{
\method{set_params}{reg_knn}(obj, params)
}
\value{
The reg_knn object updated with the new parameter values
}
\description{
It receives as input a ts_rf object (obj) and a set of parameters (params)
}
\details{
If the parameter set contains an entry for nodesize, the corresponding value is assigned to the ts_rf object. Likewise, if the parameter set contains an entry for ntree, the corresponding value is assigned to the ts_rf object
}
