% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_mlp.R
\name{set_params.reg_mlp}
\alias{set_params.reg_mlp}
\title{Set parameters values for reg_mlp}
\usage{
\method{set_params}{reg_mlp}(obj, params)
}
\value{
The reg_mlp object updated with the new parameter values
}
\description{
It receives as input a reg_mlp object (obj) and a set of parameters (params)
}
