% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rf.R
\name{set_params.reg_rf}
\alias{set_params.reg_rf}
\title{Set parameters values for reg_rf}
\usage{
\method{set_params}{reg_rf}(obj, params)
}
\value{
The reg_rf object updated with the new parameter values
}
\description{
It receives as input a reg_rf object (obj) and a set of parameters (params)
}
