% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mlp.R
\name{set_params.ts_mlp}
\alias{set_params.ts_mlp}
\title{Update ts_mlp object parameters}
\usage{
\method{set_params}{ts_mlp}(obj, params)
}
\value{
ts_mlp object updated with new parameter values
}
\description{
It takes two arguments: the ts_mlp object to be updated (obj) and a list object containing the new parameter values to be updated (params)
}
\details{
The function checks whether each parameter is present in the params list using the is.null function. If the parameter is present, the corresponding value is assigned to the parameter in the ts_mlp object. Otherwise, the parameter is not changed
}
