% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_svm.R
\name{set_params.ts_svm}
\alias{set_params.ts_svm}
\title{Updates the parameters of an SVM model object for time series (ts_svm)}
\usage{
\method{set_params}{ts_svm}(obj, params)
}
\value{
The obj template object updated with the new parameters
}
\description{
It takes as input the obj model object and a set of params parameters to update.
}
\details{
The function checks that each parameter specified in params is not null and, if not null, updates the corresponding parameter in the object obj
}
