% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{shrink}
\alias{shrink}
\title{Time Series Augmentation Shrink}
\usage{
shrink(factor = 0.8)
}
\arguments{
\item{factor}{a real value (default = 0.8) define the degree of distortion applied.}
}
\value{
a \code{shrink} object.
}
\description{
Time Series Augmentation Shrink
}
