% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sin_data}
\alias{sin_data}
\title{Time series example dataset}
\format{
\code{data.frame}.
}
\source{
This dataset was generated for examples.
}
\usage{
data(sin_data)
}
\description{
Synthetic dataset of sine function.
\itemize{
\item x: correspond time from 0 to 10.
\item y: dependent variable for time series modeling.
}
}
\examples{
data(sin_data)
head(sin_data)
}
\keyword{datasets}
