% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{sMAPE.ts}
\alias{sMAPE.ts}
\title{Calculate the symmetric mean absolute percent error (sMAPE)}
\usage{
sMAPE.ts(actual, prediction)
}
\arguments{
\item{actual}{real observations}

\item{prediction}{predicted observations}
}
\value{
The sMAPE between the actual and prediction vectors
}
\description{
The function receives two variables as a parameter, which are actual and prediction
}
