% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{sMAPE.tsreg}
\alias{sMAPE.tsreg}
\title{Calculate the symmetric mean absolute percent error (sMAPE)}
\usage{
sMAPE.tsreg(actual, prediction)
}
\arguments{
\item{prediction}{}
}
\value{
The sMAPE between the actual and prediction vectors
}
\description{
The function receives two variables as a parameter, which are actual and prediction
}
\details{
It is an error measure that is useful for assessing the accuracy of forecasts in time series, where observations can have different scales
}
