% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_evaluation}
\alias{smoothing_evaluation}
\title{Smoothing by evaluation}
\usage{
smoothing_evaluation(data, attribute)
}
\description{
The smoothing_evaluation() function performs a smoothing evaluation using cross-validation. It calculates the mean squared error (MSE) for each smoothing parameter and returns a data frame with the results. The function uses the smoothing technique specified by the user to smooth the data.
}
\details{
The arguments of function:
data: a data frame containing the dataset to be smoothed;
target: a string indicating the target variable;
method: a string indicating the smoothing method to be used ("smoothing by interval", "smoothing by freq", "smoothing by cluster");
parameters: a vector containing the parameters to be used for the smoothing method. The number of parameters will depend on the chosen method;
k: an integer indicating the number of folds to be used for cross-validation.
}
