% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_freq}
\alias{smoothing_freq}
\title{Smoothing by Freq}
\usage{
smoothing_freq(n)
}
\description{
The 'smoothing_freq' function is used to smooth a given time series data by aggregating observations within a fixed frequency.
}
\details{
The function takes in three arguments:
ts_data: a time series object to be smoothed.
frequency: the frequency at which to aggregate the observations (e.g., "day", "week", "month").
method: the method used to calculate the smoothed values (e.g., "mean", "median", "max", "min").
}
