% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_smoothing_quantization.R
\name{smoothing_quantization}
\alias{smoothing_quantization}
\title{Smoothing by quantization (k-means)}
\usage{
smoothing_quantization(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
returns an object of class \code{smoothing_quantization}
}
\description{
Quantize a numeric vector into \code{n} levels using k‑means on the values and
replace each value by its cluster mean (vector quantization).
}
\examples{
data(iris)
obj <- smoothing_quantization(n = 2)
obj <- fit(obj, iris$Sepal.Length)
sl.bi <- transform(obj, iris$Sepal.Length)
table(sl.bi)
obj$interval

bins <- cut(iris$Sepal.Length, unique(obj$interval.adj), FALSE, include.lowest = TRUE)
entro <- evaluate(obj, bins, iris$Species)
entro$entropy
}
\references{
MacQueen, J. (1967). Some Methods for classification and Analysis of Multivariate Observations.
}
