% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{start_log}
\alias{start_log}
\title{Start logging on an obj object}
\usage{
start_log(obj)
}
\arguments{
\item{obj}{object: .}
}
\description{
It takes as parameter an obj object
}
\details{
As it makes use of UseMethod, it is expected that there are other functions that define specific behaviors for different classes of objects
}
