% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{stretch}
\alias{stretch}
\title{Time Series Augmentation Stretch}
\usage{
stretch(factor = 1.2)
}
\arguments{
\item{factor}{a real value (default = 1.2) define the degree of distortion applied.}
}
\value{
a \code{stretch} object.
}
\description{
Apply temporal distortion to the time axis of the data.
}
