% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{transform.default}
\alias{transform.default}
\title{dal_base object}
\usage{
\method{transform}{default}(obj, ...)
}
\arguments{
\item{obj, }{...}
}
\value{
It simply returns NULL, which indicates that no transforms are applied
}
\description{
A default function that defines the default behavior of the transform function for objects of class dal_transform
}
\details{
This function can be overridden by other custom transform methods for dal_transform objects that require different behavior
}
