% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{transform.minmax}
\alias{transform.minmax}
\title{Transform minmax}
\usage{
\method{transform}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{minmax object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
transformed dataset
}
\description{
Scale data using minmax normalization
}
\examples{
trans <- dal_transform()
}
