% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_outliers.R
\name{transform.outliers}
\alias{transform.outliers}
\title{Remove outliers from data}
\usage{
\method{transform}{outliers}(obj, data)
}
\arguments{
\item{data}{}
}
\value{
returns the output dataset after applying the transformation and adds an "idx" attribute that contains a boolean vector indicating which rows were removed from the original dataset
}
\description{
The function takes two arguments, the first is an outliers object returned by the outliers() function, which contains the necessary parameters to perform the transformation, including the cutoff limits calculated in the fit.outliers function. The second argument is the set of input data you want to transform
}
\details{
Removes lines that contain values that are considered outliers according to the limits calculated in the fit.outliers function
}
