% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{transform.smoothing}
\alias{transform.smoothing}
\title{Transform smoothing}
\usage{
\method{transform}{smoothing}(obj, data)
}
\arguments{
\item{obj}{object: .}

\item{data}{optional arguments./ further arguments passed to or from other methods.}
}
\description{
Transform smoothing
}
\examples{
x <- transform(obj, data)
}
