% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{transform.ts_an}
\alias{transform.ts_an}
\title{Normalize the input data}
\usage{
\method{transform}{ts_an}(obj, data, x = NULL)
}
\arguments{
\item{x}{}
}
\value{
Normalized data along with annual moving average value
}
\description{
It takes 3 arguments: obj, data and x
}
\details{
The function subtracts the value of the annual moving average (calculated by the ma.ts_an function) from each sample and then applying min-max normalization, transforming the data into a scale between 0 and 1
}
