% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{transform.ts_diff}
\alias{transform.ts_diff}
\title{Transform the data}
\usage{
\method{transform}{ts_diff}(obj, data, x = NULL)
}
\value{
Transformed full data matrix
}
\description{
It takes as parameters the variables obj, data and x
}
\details{
If x is provided, the function calculates the corresponding transformation on the time series x using the reference value (ref) and scale (sw) stored in the data attributes. If x is not provided, the function calculates the corresponding transformation on all data from the data time series, which must be a matrix with the calculated differences between the column values of the original time series
}
