% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{transform.ts_gminmax}
\alias{transform.ts_gminmax}
\title{Normalize time series data}
\usage{
\method{transform}{ts_gminmax}(obj, data, x = NULL)
}
\value{
The function returns the vector x if the argument x is given, or it returns the data matrix with the values transformed by the object obj
}
\description{
It takes as parameters the variables obj, data and x
}
\details{
Normalization is done using the "min-max scaling" technique, which involves subtracting the minimum value of the time series (obj$gmin) and dividing by the interval between the minimum and maximum value (obj$gmax-obj$gmin)
}
