% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{transform.ts_swminmax}
\alias{transform.ts_swminmax}
\title{Normalizes each line of the input data to the range \link{0,1}}
\usage{
\method{transform}{ts_swminmax}(obj, data, x = NULL)
}
\value{
If the parameter x is given, the function returns the transformation applied to x. Otherwise, the function returns the data transformed input data
}
\description{
It takes as parameters the variables obj, data, x
}
\details{
Applies the SWMinMax transformation on the input data using the normalization parameters contained in the obj object
}
