% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_zscore.R
\name{transform.zscore}
\alias{transform.zscore}
\title{transform z-score normalization}
\usage{
\method{transform}{zscore}(obj, data, ...)
}
\arguments{
\item{obj}{z-score transformation object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
transformed dataset
}
\description{
Scale data using z-score normalization
}
\examples{
# see ?zscore
}
