% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{ts_as_matrix}
\alias{ts_as_matrix}
\title{Prepare the data to feed a machine learning model}
\usage{
ts_as_matrix(data, input_size)
}
\value{
An array of the latest input_size observations of each time series at date
}
\description{
Takes a time series dataset data and an input size input_size
}
