% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_awareness.R
\name{ts_aug_awareness}
\alias{ts_aug_awareness}
\title{Time Series Awareness}
\usage{
ts_aug_awareness(factor = 1)
}
\arguments{
\item{factor}{factor of awareness}
}
\value{
a \code{ts_aug_awareness} object.
}
\description{
Time Series Awareness filter
}
\examples{
trans <- dal_transform()
}
