% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_stretch.R
\name{ts_aug_stretch}
\alias{ts_aug_stretch}
\title{Time Series Augmentation ts_aug_stretch}
\usage{
ts_aug_stretch(factor = 1.2)
}
\arguments{
\item{factor}{a real value (default = 1.2) define the degree of distortion applied.}
}
\value{
a \code{ts_aug_stretch} object.
}
\description{
Augmentation is a technique used to increase the size and
diversity of a time series dataset by creating new instances of the original
data through transformations or modifications. The goal is to improve the
performance of machine learning models trained on time series data by
reducing overfitting and improving generalization.Apply temporal distortion to the time axis of the data.
}
\examples{
trans <- dal_transform()
}
