% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_awaresmooth.R
\name{ts_augawaresmooth}
\alias{ts_augawaresmooth}
\title{Time Series Awareness Smooth}
\usage{
ts_augawaresmooth(factor = 0)
}
\arguments{
\item{factor}{factor of awareness}
}
\value{
a \code{ts_augawaresmooth}
}
\description{
Time Series Awareness Smooth filter
}
\examples{
trans <- dal_transform()
}
