% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_jitter.R
\name{ts_augjitter}
\alias{ts_augjitter}
\title{Time Series Augmentation ts_augjitter}
\usage{
ts_augjitter()
}
\value{
a \code{ts_augjitter} object
}
\description{
Augmentation is a technique used to increase the size and
diversity of a time series dataset by creating new instances of the original
data through transformations or modifications. The goal is to improve the
performance of machine learning models trained on time series data by
reducing overfitting and improving generalization. ts_augjitter adds random "noise" to each data point in the time series.
}
\examples{
trans <- dal_transform()
}
