% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_shrink.R
\name{ts_augshrink}
\alias{ts_augshrink}
\title{Time Series Augmentation ts_augshrink}
\usage{
ts_augshrink(factor = 0.8)
}
\arguments{
\item{factor}{a real value (default = 0.8) define the degree of distortion applied.}
}
\value{
a \code{ts_augshrink} object.
}
\description{
Augmentation is a technique used to increase the size and
diversity of a time series dataset by creating new instances of the original
data through transformations or modifications. The goal is to improve the
performance of machine learning models trained on time series data by
reducing overfitting and improving generalization.Time Series Augmentation ts_augshrink
}
\examples{
trans <- dal_transform()
}
