% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_swfilter.R
\name{ts_aware_smooth}
\alias{ts_aware_smooth}
\title{Time Series Awareness Smooth}
\usage{
ts_aware_smooth(factor = 0)
}
\value{
a \code{ts_aware_smooth} object.
}
\description{
Time Series Awareness Smooth
}
