% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_conv1d.R
\name{ts_conv1d}
\alias{ts_conv1d}
\title{Conv1D}
\usage{
ts_conv1d(preprocess = NA, input_size = NA, epochs = 10000L)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}

\item{epochs}{maximum number of epochs}
}
\value{
a \code{ts_conv1d} object.
}
\description{
Creates a time series prediction object that uses the Conv1D.
It wraps the pytorch library.
}
\examples{
#Use the same example of ts_mlp changing the constructor to:
model <- ts_conv1d(ts_norm_gminmax(), input_size=4, epochs = 10000L)
}
