% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_data}
\alias{ts_data}
\title{Time Series Data}
\usage{
ts_data(y, sw = 1)
}
\arguments{
\item{sw}{integer: sliding window size.}
}
\value{
a \code{ts_data} object.
}
\description{
receives a vector or matrix y containing the data of a time series and an integer sw indicating the size of the sliding window
}
\details{
Prepare data for time series analysis
}
