% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_filter.R
\name{ts_ema}
\alias{ts_ema}
\title{Time Series Exponential Moving Average}
\usage{
ts_ema(ema = 3)
}
\value{
a \code{ts_ema} object.
}
\description{
Used to smooth out fluctuations, while giving more weight to
recent observations. Particularly useful when the data has a trend or
seasonality component.
}
