% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_smooth.R
\name{ts_fil_smooth}
\alias{ts_fil_smooth}
\title{Time Series Smooth}
\usage{
ts_fil_smooth()
}
\value{
a \code{ts_fil_smooth} object.
}
\description{
Used to remove or reduce randomness (noise).
}
\examples{
trans <- dal_transform()
}
