% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_lstm.R
\name{ts_lstm}
\alias{ts_lstm}
\title{LSTM}
\usage{
ts_lstm(preprocess = NA, input_size = NA, epochs = 10000L)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}

\item{epochs}{maximum number of epochs}
}
\value{
a \code{ts_lstm} object.
}
\description{
Creates a time series prediction object that uses the LSTM.
It wraps the pytorch library.
}
\examples{
#Use the same example of ts_mlp changing the constructor to:
model <- ts_lstm(ts_norm_gminmax(), input_size=4, epochs = 10000L)
}
