% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_tune.R
\name{ts_maintune}
\alias{ts_maintune}
\title{Time Series Main Tune}
\usage{
ts_maintune(
  preprocess,
  input_size,
  base_model,
  augment = ts_augment(),
  folds = 10
)
}
\value{
a \code{ts_maintune} object.
}
\description{
Time Series Main Tune
}
