% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_none.R
\name{ts_norm_none}
\alias{ts_norm_none}
\title{no normalization}
\usage{
ts_norm_none()
}
\value{
a \code{ts_norm_none} object.
}
\description{
Does not make data normalization.
}
\examples{
library(daltoolbox)
data(sin_data)

#convert to sliding windows
xw <- ts_data(sin_data$y, 10)

#no data normalization
normalize <- ts_norm_none()
normalize <- fit(normalize, xw)
xa <- transform(normalize, xw)
ts_head(xa)
}
