% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{ts_plot}
\alias{ts_plot}
\title{Plot a time series chart}
\usage{
ts_plot(x = NULL, y, color = "black", label_x = "", label_y = "")
}
\arguments{
\item{x}{input variable}

\item{y}{output variable}

\item{color}{color for time series}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}
}
\description{
The function receives six variables as a parameter, which are obj and y, yadj, main and xlabels. The graph is plotted with 3 lines: the original series (in black), the adjusted series (in blue) and the predicted series (in green)
}
