% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{ts_plot_pred}
\alias{ts_plot_pred}
\title{Plot a time series chart}
\usage{
ts_plot_pred(
  x = NULL,
  y,
  yadj,
  ypred,
  color = "black",
  label_x = "",
  label_y = ""
)
}
\arguments{
\item{label_y}{}
}
\description{
The function receives six variables as a parameter, which are obj and y, yadj, main and xlabels. The graph is plotted with 3 lines: the original series (in black), the adjusted series (in blue) and the predicted series (in green)
}
\details{
It displays the value of the sMAPE metric for the training and test sets. The obj argument is a tsreg object that contains the information of the model used, y is the original series, yadj is the adjusted series and ypre is the predicted series
}
