% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_projection}
\alias{ts_projection}
\title{Time Series Projection}
\usage{
ts_projection(ts)
}
\arguments{
\item{ts}{matrix or data.frame containing the time series.}
}
\value{
a \code{ts_projection} object.
}
\description{
It takes a time series as input (an object of type matrix or data.frame)
}
\details{
If the time series has more than one column, the function assumes that the last column is the output variable (or response variable), and all other columns are input variables (or independent variables)
}
