% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{ts_reg}
\alias{ts_reg}
\title{Create an object of class "ts_reg" which is a subclass of object dal_learner}
\usage{
ts_reg()
}
\value{
An object of the class "ts_reg"
}
\description{
The function sets some object attributes, such as log, debug and reproduce, to TRUE or FALSE. Then it adds the "ts_reg" class to the object and returns the created object
}
\examples{
trans <- dal_transform()
}
