% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_regsw.R
\name{ts_regsw}
\alias{ts_regsw}
\title{Create an object of class "ts_regsw", which is an extension of class "ts_reg"}
\usage{
ts_regsw(preprocess = NA, input_size = NA)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}
}
\value{
A ts_reg object
}
\description{
Preprocessing parameters and input size are user specified
}
\examples{
trans <- dal_transform()
}
