% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_swminmax.R
\name{ts_swminmax}
\alias{ts_swminmax}
\title{Time Series Sliding Window Min-Max}
\usage{
ts_swminmax(remove_outliers = TRUE)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_swminmax} object.
}
\description{
It takes as parameter the variable remove_outliers. The ts_swminmax function creates an object for normalizing a time series based on the "sliding window min-max scaling" method
}
\examples{
trans <- dal_transform()
}
