% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transform.R
\name{ts_transform}
\alias{ts_transform}
\title{Time Series Transformation}
\usage{
ts_transform()
}
\value{
a \code{ts_transform} object.
}
\description{
Manipulate and reshapa a time series to better understand and
analyze the underlying patterns, trends, and relationships.
}
\examples{
trans <- dal_transform()
}
