% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_tune.R
\name{ts_tune}
\alias{ts_tune}
\title{Time Series Tune}
\usage{
ts_tune(input_size, base_model, folds = 10)
}
\arguments{
\item{input_size}{input size for machine learning model}

\item{base_model}{base model for tuning}

\item{folds}{number of folds for cross-validation}
}
\value{
a \code{ts_tune} object.
}
\description{
Time Series Tune
}
\examples{
trans <- dal_transform()
}
