% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_swfilter.R
\name{tsaug_aware_smooth}
\alias{tsaug_aware_smooth}
\title{Time Series Awareness Smooth}
\usage{
tsaug_aware_smooth(factor = 0)
}
\value{
a \code{tsaug_aware_smooth} object.data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAWElEQVR42mNgGPTAxsZmJsVqQApgmGw1yApwKcQiT7phRBuCzzCSDSHGMKINIeDNmWQlA2IigKJwIssQkHdINgxfmBBtGDEBS3KCxBc7pMQgMYE5c/AXPwAwSX4lV3pTWwAAAABJRU5ErkJggg==
}
\description{
Time Series Awareness Smooth
}
