% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{tsaug_flip}
\alias{tsaug_flip}
\title{Time Series Augmentation tsaug_flip}
\usage{
tsaug_flip()
}
\value{
a \code{tsaug_flip} object.
}
\description{
reverse the order of the data along the time axis.
}
