% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{tsaug_jitter}
\alias{tsaug_jitter}
\title{Time Series Augmentation tsaug_jitter}
\usage{
tsaug_jitter()
}
\value{
a \code{tsaug_jitter} object
}
\description{
tsaug_jitter adds random "noise" to each data point in the time series.
}
\examples{
trans <- dal_transform()
}
