% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{tsaug_shrink}
\alias{tsaug_shrink}
\title{Time Series Augmentation tsaug_shrink}
\usage{
tsaug_shrink(factor = 0.8)
}
\arguments{
\item{factor}{a real value (default = 0.8) define the degree of distortion applied.}
}
\value{
a \code{tsaug_shrink} object.
}
\description{
Time Series Augmentation tsaug_shrink
}
\examples{
trans <- dal_transform()
}
