% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{tsaug_stretch}
\alias{tsaug_stretch}
\title{Time Series Augmentation tsaug_stretch}
\usage{
tsaug_stretch(factor = 1.2)
}
\arguments{
\item{factor}{a real value (default = 1.2) define the degree of distortion applied.}
}
\value{
a \code{tsaug_stretch} object.
}
\description{
Apply temporal distortion to the time axis of the data.
}
\examples{
trans <- dal_transform()
}
