% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_augment.R
\name{tsaug_wormhole}
\alias{tsaug_wormhole}
\title{Time Series Augmentation tsaug_wormhole}
\usage{
tsaug_wormhole()
}
\value{
a \code{tsaug_wormhole} object.
}
\description{
Inserts or removes segments of the time series data.
}
