% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_ema.R
\name{tsfil_ema}
\alias{tsfil_ema}
\title{Time Series Exponential Moving Average}
\usage{
tsfil_ema(ema = 3)
}
\arguments{
\item{ema}{exponential moving average size}
}
\value{
a \code{tsfil_ema} object.
}
\description{
Used to smooth out fluctuations, while giving more weight to
recent observations. Particularly useful when the data has a trend or
seasonality component.
}
\examples{
trans <- dal_transform()
}
