% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_filter.R
\name{tsfil_filter}
\alias{tsfil_filter}
\title{Time Series Filter}
\usage{
tsfil_filter()
}
\value{
a \code{tsfil_filter} object.
}
\description{
Used to extract or remove specific components from a time series.
}
