% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_filter.R
\name{tsfil_smooth}
\alias{tsfil_smooth}
\title{Time Series Smooth}
\usage{
tsfil_smooth()
}
\value{
a \code{tsfil_smooth} object.
}
\description{
Used to remove or reduce randomness (noise).
}
\examples{
trans <- dal_transform()
}
