% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{tshead}
\alias{tshead}
\title{Extract the first observations from a time series}
\usage{
tshead(x, n = 6L, ...)
}
\arguments{
\item{x}{input variable}

\item{n}{integer: size of test data.}

\item{...}{optional arguments}
}
\value{
The first n observations of a time series x
}
\description{
The function takes as arguments the variables x, n (default = 6L), ...
}
\examples{
trans <- dal_transform()
}
