% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{tsplot}
\alias{tsplot}
\title{Plot a time series chart}
\usage{
tsplot(obj, y, yadj, ypre, main = NULL, xlabels = NULL)
}
\arguments{
\item{xlabels}{}
}
\description{
The function receives six variables as a parameter, which are obj and y, yadj, main and xlabels. The graph is plotted with 3 lines: the original series (in black), the adjusted series (in blue) and the predicted series (in green)
}
\details{
It displays the value of the sMAPE metric for the training and test sets. The obj argument is a tsreg object that contains the information of the model used, y is the original series, yadj is the adjusted series and ypre is the predicted series
}
