% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{tsreg}
\alias{tsreg}
\title{Create an object of class "tsreg" which is a subclass of object "dal_transform}
\usage{
tsreg()
}
\value{
An object of the class "tsreg"
}
\description{
The function sets some object attributes, such as log, debug and reproduce, to TRUE or FALSE. Then it adds the "tsreg" class to the object and returns the created object
}
\examples{
trans <- dal_transform()
}
