% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_arima.R
\name{tsreg_arima}
\alias{tsreg_arima}
\title{Time Series AutoRegressive Integrated Moving Average (ARIMA)}
\usage{
tsreg_arima()
}
\value{
obj
}
\description{
ARIMA is a regression model for forecasting time series data
}
\examples{
trans <- dal_transform()
}
