% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{tsreg_sw}
\alias{tsreg_sw}
\title{Create an object of class "tsreg_sw", which is an extension of class "tsreg"}
\usage{
tsreg_sw(preprocess = NA, input_size = NA)
}
\value{
A tsreg object
}
\description{
Preprocessing parameters and input size are user specified
}
