% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_vae_encode.R
\name{vae_encode}
\alias{vae_encode}
\title{Variational Autoencoder - Encode}
\usage{
vae_encode(
  input_size,
  encoding_size,
  mean_var_size = 6,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{mean_var_size}{mean variable size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
a \code{vae_encode_decode} object.
}
\description{
Creates an deep learning variational autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See example at https://nbviewer.org/github/cefet-rj-dal/daltoolbox-examples
}
