% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_varae_encode.R
\name{varae_encode}
\alias{varae_encode}
\title{Variational Autoencoder - Encode}
\usage{
varae_encode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}
}
\value{
returns a \code{varae_encode} object.
}
\description{
Creates an deep learning variational autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See an example of using `varae_encode` at this
#[link](https://github.com/cefet-rj-dal/daltoolbox/blob/main/transf/varae_encode.ipynb)
}
