% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_zscore.R
\name{zscore}
\alias{zscore}
\title{z-score normalization}
\usage{
zscore(nmean = 0, nsd = 1)
}
\arguments{
\item{nmean}{mean}

\item{nsd}{standard deviation}
}
\value{
z-score transformation object
}
\description{
Scale data using z-score normalization
}
\examples{
data(iris)
trans <- zscore()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
}
