% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{action.default}
\alias{action.default}
\title{Default Action implementation}
\usage{
\method{action}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
It simply returns NULL, which indicates that no transforms are applied
}
\description{
A default function that defines the default behavior of the transform function for objects of class dal_transform
}
\examples{
trans <- dal_transform()
}
