% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{action}
\alias{action}
\title{Action}
\usage{
action(obj, ...)
}
\arguments{
\item{obj}{object: a dal_base object to apply the transformation on the input dataset.}

\item{...}{optional arguments.}
}
\value{
returns the result of an action of the model applied in provided data
}
\description{
Executes the action of model applied in provided data
}
\examples{
data(iris)
# an example is minmax normalization
trans <- minmax()
trans <- fit(trans, iris)
tiris <- action(trans, iris)
}
