% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_class_label}
\alias{adjust_class_label}
\title{compute categorical mapping}
\usage{
adjust_class_label(x, valTrue = 1, valFalse = 0)
}
\arguments{
\item{x}{vector to be categorized}

\item{valTrue}{\itemize{
\item value to represent true
}}

\item{valFalse}{\itemize{
\item value to represent false
}}
}
\value{
factor
}
\description{
compute categorical mapping
}
\examples{
trans <- dal_transform()
}
