% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjust.R
\name{adjust_data.frame}
\alias{adjust_data.frame}
\title{Adjust to data frame}
\usage{
adjust_data.frame(data)
}
\arguments{
\item{data}{dataset}
}
\value{
returns a data.frame
}
\description{
Coerce an object to \code{data.frame} if needed (useful for S3 methods in this package).
}
\examples{
data(iris)
df <- adjust_data.frame(iris)
}
