% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{adjust_data.frame}
\alias{adjust_data.frame}
\title{dal_base object}
\usage{
adjust_data.frame(data)
}
\value{
The date argument
}
\description{
It takes as parameter an obj object
}
\details{
This verification is useful when object is not a dataframe
}
