% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{adjust_data.frame}
\alias{adjust_data.frame}
\title{dal_base object}
\usage{
adjust_data.frame(data)
}
\arguments{
\item{data}{dataset}
}
\value{
The date argument
}
\description{
It takes as parameter an obj object
}
\examples{
trans <- dal_transform()
}
