% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_data.frame}
\alias{adjust_data.frame}
\title{Adjust to data frame}
\usage{
adjust_data.frame(data)
}
\arguments{
\item{data}{dataset}
}
\value{
The date argument
}
\description{
dataset data is adjusted to a \code{data.frame}
}
\examples{
data(iris)
df <- adjust_data.frame(iris)
}
