% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjust.R
\name{adjust_factor}
\alias{adjust_factor}
\title{Adjust factors}
\usage{
adjust_factor(value, ilevels, slevels)
}
\arguments{
\item{value}{vector to be converted into factor}

\item{ilevels}{order for categorical values}

\item{slevels}{labels for categorical values}
}
\value{
returns an adjusted factor
}
\description{
Convert a vector to a factor with specified internal levels (\code{ilevels}) and labels (\code{slevels}).
}
\details{
Numeric vectors are first converted to factors with \code{ilevels} as the level order, then relabeled to \code{slevels}.
}
