% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_factor}
\alias{adjust_factor}
\title{adjust factors}
\usage{
adjust_factor(value, ilevels, slevels)
}
\arguments{
\item{value}{vector to be converted into factor}

\item{ilevels}{order for categorical values}

\item{slevels}{labels for categorical values}
}
\value{
an adjusted factor
}
\description{
vector \code{value} is adjusted to a factor
}
