% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjust.R
\name{adjust_matrix}
\alias{adjust_matrix}
\title{Adjust to matrix}
\usage{
adjust_matrix(data)
}
\arguments{
\item{data}{dataset}
}
\value{
returns an adjusted matrix
}
\description{
Converts a dataset to a matrix format if it is not already in that format
}
\examples{
data(iris)
mat <- adjust_matrix(iris)
}
