% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_matrix}
\alias{adjust_matrix}
\title{Adjust input data to matrix}
\usage{
adjust_matrix(data)
}
\arguments{
\item{data}{dataset}
}
\description{
adjust_matrix function receives
a data object and converts it to a matrix.

It returns the same input if it is already a matrix.
}
\examples{
data(iris)
mat <- adjust_matrix(iris)
}
