% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{adjust_matrix}
\alias{adjust_matrix}
\title{dal_base object}
\usage{
adjust_matrix(data)
}
\arguments{
\item{data}{dataset}
}
\value{
An array corresponding to the object passed as a parameter, if it is not an array. If the object is already an array, the function simply returns it
}
\description{
It takes a data object as an argument
}
\examples{
trans <- dal_transform()
}
