% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjust.R
\name{adjust_ts_data}
\alias{adjust_ts_data}
\title{Adjust \code{ts_data}}
\usage{
adjust_ts_data(data)
}
\arguments{
\item{data}{dataset}
}
\value{
returns an adjusted \code{ts_data}
}
\description{
Converts a dataset to a \code{ts_data} object
}
