% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_adjusts.R
\name{adjust_ts_data}
\alias{adjust_ts_data}
\title{adjust \code{ts_data}}
\usage{
adjust_ts_data(data)
}
\arguments{
\item{data}{dataset}
}
\value{
an adjusted \code{ts_data}
}
\description{
dataset data is adjusted to a \code{ts_data}
}
