% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_bal_subsampling.R
\name{bal_subsampling}
\alias{bal_subsampling}
\title{Random class undersampling}
\usage{
bal_subsampling(attribute, seed = NULL)
}
\arguments{
\item{attribute}{target class attribute name}

\item{seed}{optional random seed for reproducibility}
}
\value{
returns an object of class \code{bal_subsampling}
}
\description{
Balance class distributions by randomly reducing all classes to the minority count.
}
\examples{
data(iris)
iris_imb <- iris[c(1:50, 51:71, 101:111), ]
bal <- bal_subsampling("Species", seed = 123)
iris_bal <- transform(bal, iris_imb)
table(iris_bal$Species)
}
