% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_dtree.R
\name{cla_dtree}
\alias{cla_dtree}
\title{Decision Tree Classification}
\usage{
cla_dtree(attribute, slevels)
}
\arguments{
\item{attribute}{attribute target to model building.}

\item{slevels}{The possible values for the target classification.}
}
\value{
A classification object that uses the Decision Tree algorithm for classification.
}
\description{
Creates a classification object that uses the Decision Tree algorithm for classification.
}
\examples{
trans <- dal_transform()
}
