% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_knn.R
\name{cla_knn}
\alias{cla_knn}
\title{K Nearest Neighbor Classification}
\usage{
cla_knn(attribute, slevels = NULL, k = 1:30)
}
\arguments{
\item{attribute}{Name of the attribute used as target classification.}

\item{slevels}{Possible values for the target classification.}

\item{k}{A vector of integers indicating the number of neighbors to be considered.}
}
\value{
A classification object.
}
\description{
Classifies using the K-Nearest Neighbor algorithm.
}
\details{
This function implements the K-Nearest Neighbor algorithm for classification. It computes the K nearest neighbors of each data point in the training set, and then determines the class label of each point by taking the majority vote of the K neighbors.
}
