% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_majority.R
\name{cla_majority}
\alias{cla_majority}
\title{Majority Classification}
\usage{
cla_majority(attribute, slevels)
}
\arguments{
\item{attribute}{Name of the attribute used as target classification.}

\item{slevels}{Possible values for the target classification.}
}
\value{
Returns a classification object.
}
\description{
This function creates a classification object that uses the majority vote strategy to predict the target attribute. Given a target attribute, the function counts the number of occurrences of each value in the dataset and selects the one that appears most often.
}
\details{
The function takes two arguments: the name of the attribute used as target classification and the possible values for the target classification. The second argument is optional and should be provided only if the dataset does not contain all possible values for the target attribute.
}
