% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_mlp.R
\name{cla_mlp}
\alias{cla_mlp}
\title{Classification using Artificial Neural Network (ANN)}
\usage{
cla_mlp(attribute, slevels, size = NULL, decay = 0.1, maxit = 1000)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{possible values for the target classification}

\item{size}{number of nodes that will be used in the hidden layer}

\item{decay}{how quickly it decreases in gradient descent}

\item{maxit}{maximum iterations}
}
\value{
a classification object
}
\description{
Performs classification using Artificial Neural Network (ANN) algorithm
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_mlp("Species", slevels, size=3, decay=0.03)

# preparing dataset for random sampling
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, iris)
iris_train <- sr$train
iris_test <- sr$test

model <- fit(model, iris_train)

prediction <- predict(model, iris_test)
predictand <- adjust_class_label(iris_test[,"Species"])
train_eval <- evaluate(model, predictand, prediction)
train_eval$metrics
}
