% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_nb.R
\name{cla_nb}
\alias{cla_nb}
\title{Naive Bayes Classifier}
\usage{
cla_nb(attribute, slevels)
}
\arguments{
\item{attribute}{attribute target to model building.}

\item{slevels}{Possible values for the target classification.}
}
\value{
A classification object.
}
\description{
Classification using the Naive Bayes algorithm
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_nb("Species", slevels)

# preparing dataset for random sampling
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, iris)
iris_train <- sr$train
iris_test <- sr$test

model <- fit(model, iris_train)

prediction <- predict(model, iris_test)
predictand <- adjust_class_label(iris_test[,"Species"])
train_eval <- evaluate(model, predictand, prediction)
train_eval$metrics
}
