% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_nb.R
\name{cla_nb}
\alias{cla_nb}
\title{Naive Bayes Classifier}
\usage{
cla_nb(attribute, slevels)
}
\arguments{
\item{attribute}{attribute target to model building.}

\item{slevels}{Possible values for the target classification.}
}
\value{
A classification object.
}
\description{
Classification using the Naive Bayes algorithm
}
\examples{
trans <- dal_transform()
}
