% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{cla_rf}
\alias{cla_rf}
\title{Random Forest Classifier}
\usage{
cla_rf(attribute, slevels, mtry = NULL, ntree = 10)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{mtry}{number of attributes to build tree}

\item{ntree}{number of trees}
}
\value{
obj
}
\description{
Random Forest Classifier
}
\examples{
trans <- dal_transform()
}
