% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_rf.R
\name{cla_rf}
\alias{cla_rf}
\title{Random Forest Classifier}
\usage{
cla_rf(attribute, slevels, mtry = NULL, ntree = 10)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{mtry}{number of attributes to build tree}

\item{ntree}{number of trees}
}
\value{
obj
}
\description{
Random Forest Classifier
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_rf("Species", slevels, mtry=3, ntree=5)

# preparing dataset for random sampling
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, iris)
train <- sr$train
test <- sr$test

model <- fit(model, train)

prediction <- predict(model, test)
predictand <- adjust_class_label(test[,"Species"])
test_eval <- evaluate(model, predictand, prediction)
test_eval$metrics
}
