% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_svm.R
\name{cla_svm}
\alias{cla_svm}
\title{Support Vector Machine Classification}
\usage{
cla_svm(attribute, slevels, epsilon = 0.1, cost = 10, kernel = "radial")
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{epsilon}{\itemize{
\item parameter that controls the width of the margin around the separating hyperplane
}}

\item{cost}{\itemize{
\item parameter that controls the trade-off between having a wide margin and correctly classifying training data points
}}

\item{kernel}{\itemize{
\item the type of kernel function to be used in the SVM algorithm (linear, radial, polynomial, sigmoid)
}}
}
\value{
classification object
}
\description{
Classification using Support Vector Machine (SVM) algorithm
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_svm("Species", slevels, epsilon=0.0,cost=20.000)

# preparing dataset for random sampling
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, iris)
iris_train <- sr$train
iris_test <- sr$test

model <- fit(model, iris_train)

prediction <- predict(model, iris_test)
predictand <- adjust_class_label(iris_test[,"Species"])
train_eval <- evaluate(model, predictand, prediction)
train_eval$metrics
}
