% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{cla_tune}
\alias{cla_tune}
\title{Regression Tune}
\usage{
cla_tune(base_model, folds = 10, metric = "accuracy")
}
\arguments{
\item{base_model}{base model for tuning}

\item{folds}{number of folds for cross-validation}

\item{metric}{metric used to optimize}
}
\value{
a \code{cla_tune} object.
}
\description{
Regression Tune
}
\examples{
# preparing dataset for random sampling
set.seed(1)
sr <- sample_random()
sr <- train_test(sr, iris)
iris_train <- sr$train
iris_test <- sr$test

# hyper parameter optimization
tune <- cla_tune(cla_svm("Species", levels(iris$Species)))
ranges <- list(epsilon=seq(0,1,0.25), cost=seq(25,100,25), kernel = c("radial"))
model <- fit(tune, iris_train, ranges)
train_prediction <- predict(model, iris_train)

# testing optimization
test_prediction <- predict(model, iris_test)
test_predictand <- adjust_class_label(iris_test[,"Species"])
test_eval <- evaluate(model, test_predictand, test_prediction)
test_eval$metrics
}
