% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_classification.R
\name{classification}
\alias{classification}
\title{classification}
\usage{
classification(attribute, slevels)
}
\arguments{
\item{attribute}{attribute target to model building}

\item{slevels}{possible values for the target classification}
}
\value{
returns a classification object
}
\description{
Ancestor class for classification problems using MLmetrics nnet
}
\examples{
#See ?cla_dtree for a classification example using a decision tree
}
