% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_tune.R
\name{clu_tune}
\alias{clu_tune}
\title{Clustering Tune}
\usage{
clu_tune(base_model)
}
\arguments{
\item{base_model}{base model for tuning}
}
\value{
returns a \code{clu_tune} object.
}
\description{
Creates an object for tuning clustering models.
This object can be used to fit and optimize clustering algorithms by specifying hyperparameter ranges
}
\examples{
data(iris)

# fit model
model <- clu_tune(cluster_kmeans(k = 0))
ranges <- list(k = 1:10)
model <- fit(model, iris[,1:4], ranges)
model$k
}
