% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_clusterer.R
\name{cluster}
\alias{cluster}
\title{Cluster}
\usage{
cluster(obj, ...)
}
\arguments{
\item{obj}{a \code{clusterer} object}

\item{...}{optional arguments}
}
\value{
clustered data
}
\description{
Generic for clustering methods
}
\examples{
#See ?cluster_kmeans for an example of transformation
}
