% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_main.R
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(obj, ...)
}
\arguments{
\item{obj}{object: .}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Models a time series dataset by estimating the underlying trend
and seasonality components. Used to make predictions and forecast future
values of the time series based on the historical data.
}
