% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_kmeans.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{Clustering using k-means}
\usage{
cluster_kmeans(k)
}
\arguments{
\item{k}{\itemize{
\item number of clusters
}}
}
\value{
clustering object
}
\description{
Clustering using k-means
}
