% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_kmeans.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{Clustering using k-means}
\usage{
cluster_kmeans(k)
}
\arguments{
\item{k}{The number of clusters to form.}
}
\value{
A clusterer object.
}
\description{
This function applies the k-means clustering algorithm to a given data set and returns a clusterer object.
}
\examples{
trans <- dal_transform()
}
