% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_pam.R
\name{cluster_pam}
\alias{cluster_pam}
\title{Partition Around Medoids (PAM) Clustering}
\usage{
cluster_pam(k)
}
\arguments{
\item{k}{The number of clusters to generate.}
}
\value{
A clustering object.
}
\description{
Implements the Partition Around Medoids (PAM) clustering algorithm.
PAM is a partitioning clustering method that seeks to minimize the sum of dissimilarities between data points and the medoid of its assigned cluster. Unlike k-means, PAM can handle categorical data as well as distance measures other than Euclidean distance.
}
\details{
The algorithm selects k medoids (representative points) from the data set, and assigns each data point to the closest medoid. It then tries to improve the solution by swapping a medoid with a non-medoid point and computing the total dissimilarity for the new configuration. If the total dissimilarity is reduced by the swap, the solution is updated; otherwise, the swap is rejected.
}
