% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_pam.R
\name{cluster_pam}
\alias{cluster_pam}
\title{Partition Around Medoids (PAM) Clustering}
\usage{
cluster_pam(k)
}
\arguments{
\item{k}{The number of clusters to generate.}
}
\value{
A clusterer object.
}
\description{
Implements the Partition Around Medoids (PAM) clustering algorithm.
PAM is a partitioning clustering method that seeks to minimize the sum of dissimilarities between data points and the medoid of its assigned cluster. Unlike k-means, PAM can handle categorical data as well as distance measures other than Euclidean distance.
}
\examples{
trans <- dal_transform()
}
