% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_pam.R
\name{cluster_pam}
\alias{cluster_pam}
\title{Partition Around Medoids (PAM) Clustering}
\usage{
cluster_pam(k)
}
\arguments{
\item{k}{\itemize{
\item number of clusters
}}
}
\value{
clustering object
}
\description{
Partition Around Medoids (PAM) Clustering
}
