% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_clusterer.R
\name{clusterer}
\alias{clusterer}
\title{Clusterer}
\usage{
clusterer()
}
\value{
returns a \code{clusterer} object
}
\description{
Ancestor class for clustering problems
}
\examples{
#See ?cluster_kmeans for an example of transformation
}
