% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_dns_encode.R
\name{dae_encode}
\alias{dae_encode}
\title{Denoising Autoencoder - Encode}
\usage{
dae_encode(
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001,
  noise_factor = 0.3
)
}
\arguments{
\item{input_size}{input size}

\item{encoding_size}{encoding size}

\item{batch_size}{size for batch learning}

\item{num_epochs}{number of epochs for training}

\item{learning_rate}{learning rate}

\item{noise_factor}{level of noise to be added to the data}
}
\value{
a \code{dae_encode_decode} object.
}
\description{
Creates an deep learning denoising autoencoder to encode a sequence of observations.
It wraps the pytorch library.
}
\examples{
#See example at https://nbviewer.org/github/cefet-rj-dal/daltoolbox-examples
}
