% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{dal_base}
\alias{dal_base}
\title{Class dal_base}
\usage{
dal_base()
}
\value{
returns a dal_base object
}
\description{
The dal_base class is an abstract class for all dal descendants classes. It provides both fit() and action() functions
}
\examples{
trans <- dal_base()
}
