% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{dal_base}
\alias{dal_base}
\title{Class dal_base}
\usage{
dal_base()
}
\value{
returns a \code{dal_base} object
}
\description{
Minimal abstract base class for all DAL objects. Defines the common generics \code{fit()} and \code{action()}
used by transforms and learners.
}
\examples{
trans <- dal_base()
}
