% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{dal_base}
\alias{dal_base}
\title{dal_base object}
\usage{
dal_base()
}
\value{
An empty object
}
\description{
Creates a base object that will be used by other library functions
}
\details{
It creates an empty object, adds some boolean properties (log, debug and reproduce) to it, and then sets the object's class to "dal_base"
}
