% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_learner.R
\name{dal_learner}
\alias{dal_learner}
\title{DAL Learner}
\usage{
dal_learner()
}
\value{
returns a learner
}
\description{
A ancestor class for clustering, classification, regression, and time series regression.
It also provides the basis for specialized evaluation of learning performance.

An example of a learner is a decision tree (cla_dtree)
}
\examples{
#See ?cla_dtree for a classification example using a decision tree
}
