% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_learner.R
\name{dal_learner}
\alias{dal_learner}
\title{DAL Transform}
\usage{
dal_learner()
}
\value{
a dal_transform object
}
\description{
A transformation function can be applied to a time series dataset to alter its properties.
}
\examples{
trans <- dal_transform()
}
