% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_transform.R
\name{dal_transform}
\alias{dal_transform}
\title{DAL Transform}
\usage{
dal_transform()
}
\value{
returns a \code{dal_transform} object.
}
\description{
A transformation method applied to a dataset.
If needed, the fit can be called to adjust the transform.
}
\examples{
#See ?minmax for an example of transformation
}
