% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{dal_transform}
\alias{dal_transform}
\title{DAL Transform}
\usage{
dal_transform()
}
\value{
a dal_transform object
}
\description{
A transformation method applied to a dataset.
Fit can be called before transform, if needed.
}
\examples{
data(iris)
#An example of dal_transform is minmax()
trans <- minmax()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
}
