% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_tune.R
\name{dal_tune}
\alias{dal_tune}
\title{DAL Tune}
\usage{
dal_tune(base_model, folds = 10)
}
\arguments{
\item{base_model}{base model for tuning}

\item{folds}{number of folds for cross-validation}
}
\value{
returns a \code{dal_tune} object
}
\description{
Creates an ancestor class for hyperparameter optimization,
allowing the tuning of a base model using cross-validation.
}
\examples{
#See ?cla_tune for classification tuning
#See ?reg_tune for regression tuning
#See ?ts_tune for time series tuning
}
