% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample.R
\name{data_sample}
\alias{data_sample}
\title{Data Sample}
\usage{
data_sample()
}
\description{
The data_sample function in R is used to randomly sample data from a given data frame. It can be used to obtain a subset of data for further analysis or modeling.
}
\details{
The function takes the following arguments:
data: a data frame or a matrix containing the data to be sampled
size: the number of samples to be drawn from the data. If NULL, the function will return a data frame with the same number of rows as the input data
replace: a logical value indicating whether the samples should be drawn with or without replacement. The default is FALSE (sampling without replacement)
prob: a vector of probabilities for each row of the data. The length of this vector should be equal to the number of rows in the data. If NULL, all rows will have equal probability of being selected;
seed: an optional integer value that can be used to set the random number generator seed for reproducibility.
}
