% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{describe.default}
\alias{describe.default}
\title{dal_base object}
\usage{
\method{describe}{default}(obj)
}
\arguments{
\item{obj}{object}
}
\value{
description of object
}
\description{
Generates a summary object status
}
\examples{
trans <- dal_transform()
}
