% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{describe}
\alias{describe}
\title{Describe object status}
\usage{
describe(obj)
}
\arguments{
\item{obj}{object}
}
\value{
description of object
}
\description{
Generates a summary object status
}
\examples{
trans <- dal_transform()
}
