% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{describe.ts_transform}
\alias{describe.ts_transform}
\title{Describe the ts_transform function}
\usage{
\method{describe}{ts_transform}(obj)
}
\arguments{
\item{obj}{object}
}
\value{
The string "none"
}
\description{
The function receives as argument the variable obj
}
\examples{
trans <- dal_transform()
}
