% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_pattern_miner.R
\name{discover}
\alias{discover}
\title{Discover}
\usage{
discover(obj, ...)
}
\arguments{
\item{obj}{a \code{pattern_miner} object}

\item{...}{optional arguments}
}
\value{
discovered patterns
}
\description{
Generic for pattern discovery.
}
