% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_fit}
\alias{do_fit}
\title{Fit Time Series Model}
\usage{
do_fit(obj, x, y = NULL)
}
\arguments{
\item{obj}{an object representing the model or algorithm to be fitted}

\item{x}{a matrix or data.frame containing the input features for training the model}

\item{y}{a vector or matrix containing the output values to be predicted by the model}
}
\value{
returns a fitted object
}
\description{
The actual time series model fitting.
This method should be override by descendants.
}
