% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_fit}
\alias{do_fit}
\title{do fit for time series}
\usage{
do_fit(obj, x, y = NULL)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}

\item{y}{output variable}
}
\value{
fitted object
}
\description{
The actual time series model fitting.
This method should be override by descendants.
}
\examples{
trans <- dal_transform()
}
