% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_svm.R
\name{do_fit.ts_svm}
\alias{do_fit.ts_svm}
\title{Fits an SVM model to time series}
\usage{
\method{do_fit}{ts_svm}(obj, x, y)
}
\arguments{
\item{y}{}
}
\value{
The obj model object with the adjusted model
}
\description{
It takes as input the model object obj, the input data x and the expected outputs y
}
\details{
The function uses the e1071 library implementation to tune the SVM model, specifying the following parameters: x: a matrix or a data frame with the explanatory variables; y: a vector with the response variable; epsilon: The SVM smoothing parameter; cost: the SVM cost parameter; kernel: the kernel type to be used by the SVM model
}
