% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_predict}
\alias{do_predict}
\title{Predict Time Series Model}
\usage{
do_predict(obj, x)
}
\arguments{
\item{obj}{an object representing the fitted model or algorithm}

\item{x}{a matrix or data.frame containing the input features for making predictions}
}
\value{
returns the predicted values
}
\description{
The actual time series model prediction.
This method should be override by descendants.
}
