% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{do_predict}
\alias{do_predict}
\title{Define a generic method and delegate the implementation}
\usage{
do_predict(obj, x)
}
\arguments{
\item{obj}{object: .}
}
\description{
This function defines the "do_predict" method for a generic object "obj" and a dataset "x"
}
