% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_predict}
\alias{do_predict}
\title{do predict for time series}
\usage{
do_predict(obj, x)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}
}
\value{
predicted values
}
\description{
The actual time series model prediction.
This method should be override by descendants.
}
