% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_elm.R
\name{do_predict.ts_elm}
\alias{do_predict.ts_elm}
\title{Make predictions using a previously trained ELM model}
\usage{
\method{do_predict}{ts_elm}(obj, x)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}
}
\value{
A numerical object containing the predictions generated by the ELM model for the predictor variables x.
}
\description{
This function takes obj and x as arguments
}
\examples{
trans <- dal_transform()
}
