% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{do_predict.tsreg_sw}
\alias{do_predict.tsreg_sw}
\title{Performs the prediction step of the temporal regression model}
\usage{
\method{do_predict}{tsreg_sw}(obj, x)
}
\arguments{
\item{obj}{object}

\item{x}{input variable}
}
\value{
The forecast matrix
}
\description{
It receives as input the obj object that contains the trained model and the input matrix x
}
