% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_pca.R
\name{dt_pca}
\alias{dt_pca}
\title{PCA}
\usage{
dt_pca(attribute = NULL, components = NULL)
}
\description{
PCA (Principal Component Analysis) is a commonly used unsupervised dimensionality reduction technique in data analysis and machine learning. It is used to transform a dataset of possibly correlated variables into a new set of uncorrelated variables called principal components.
}
\details{
The prcomp() function returns an object of class prcomp, which contains the following components: sdev: the standard deviations of the principal components; rotation: the matrix of variable loadings (the eigenvectors of the covariance matrix); center: the centering values used, if any; scale: the scaling values used, if any; x: the centered and optionally scaled data matrix; rank: the rank of the data matrix; call: the function call used to generate the result; na.action: the method used to handle missing values, if any
}
