% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_learner.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate}
\usage{
evaluate(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
returns the evaluation
}
\description{
Evaluate learner performance.
The actual evaluate varies according to the type of learner (clustering, classification, regression, time series regression)
}
\examples{
data(iris)
slevels <- levels(iris$Species)
model <- cla_dtree("Species", slevels)
model <- fit(model, iris)
prediction <- predict(model, iris)
predictand <- adjust_class_label(iris[,"Species"])
test_eval <- evaluate(model, predictand, prediction)
test_eval$metrics
}
