% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(obj, values, prediction)
}
\arguments{
\item{obj}{object: .}

\item{values}{data set with hyper parameters.}

\item{predict}{data set with hyper parameters.}
}
\description{
evaluate
}
