% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_learner.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
evaluation
}
\description{
evaluate
}
\examples{
trans <- dal_transform()
}
