% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{fit.cla_tune}
\alias{fit.cla_tune}
\title{tune hyperparameters of ml model}
\usage{
\method{fit}{cla_tune}(obj, data, ranges, ...)
}
\arguments{
\item{obj}{an object containing the model and tuning configuration}

\item{data}{the dataset used for training and evaluation}

\item{ranges}{a list of hyperparameter ranges to explore}

\item{...}{optional arguments}
}
\value{
a fitted obj
}
\description{
Tunes the hyperparameters of a machine learning model for classification
}
