% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_dbscan.R
\name{fit.cluster_dbscan}
\alias{fit.cluster_dbscan}
\title{fit dbscan model}
\usage{
\method{fit}{cluster_dbscan}(obj, data, ...)
}
\arguments{
\item{obj}{an object containing the DBSCAN model configuration, including \code{minPts} and optionally \code{eps}}

\item{data}{the dataset to use for fitting the model}

\item{...}{optional arguments}
}
\value{
returns a fitted obj with the \code{eps} parameter set
}
\description{
Fits a DBSCAN clustering model by setting the \code{eps} parameter.
If \code{eps} is not provided, it is estimated based on the k-nearest neighbor distances.
It wraps dbscan library
}
