% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{fit.default}
\alias{fit.default}
\title{fit default implementation}
\usage{
\method{fit}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
obj
}
\description{
fit efault implementation
}
\examples{
trans <- dal_transform()
}
