% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{fit.minmax}
\alias{fit.minmax}
\title{Fit MinMax Normalization}
\usage{
\method{fit}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{minmax object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted obj
}
\description{
Compute normalization parameters
}
\examples{
trans <- dal_transform()
}
