% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_minmax.R
\name{fit.minmax}
\alias{fit.minmax}
\title{fit min-max normalization}
\usage{
\method{fit}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{min-max transformation object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted min-max transformation object
}
\description{
Compute min-max normalization parameters
}
\examples{
# see ?minmax
}
