% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{fit}
\alias{fit}
\title{Fit}
\usage{
fit(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments.}
}
\value{
obj
}
\description{
Models a time series dataset by estimating the underlying trend
and seasonality components. Used to make predictions and forecast future
values of the time series based on the historical data.
}
\examples{
trans <- dal_transform()
}
