% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{fit}
\alias{fit}
\title{Fit}
\usage{
fit(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments.}
}
\value{
obj
}
\description{
Fits a model.
}
\examples{
data(iris)
# an example is minmax normalization
trans <- minmax()
trans <- fit(trans, iris)
tiris <- action(trans, iris)
}
