% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_diff.R
\name{fit.ts_diff}
\alias{fit.ts_diff}
\title{Fit the object created by the ts_diff function to the training data}
\usage{
\method{fit}{ts_diff}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
The updated obj object
}
\description{
It takes two arguments: obj and data
}
\examples{
trans <- dal_transform()
}
