% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{fit.ts_gminmax}
\alias{fit.ts_gminmax}
\title{Sets the parameters of an object}
\usage{
\method{fit}{ts_gminmax}(obj, data)
}
\arguments{
\item{data}{}
}
\value{
The updated obj object
}
\description{
It takes 2 parameters: obj and data. This function adjusts the parameters of the obj object to the input data data using the global minimum and maximum transform (gminmax)
}
\details{
If remove_outliers is true, outliers are removed from the series before adjusting parameters. Then, the minimum and maximum values of the series after removing the outliers are stored in obj$gmin and obj$gmax, respectively
}
