% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_swminmax.R
\name{fit.ts_swminmax}
\alias{fit.ts_swminmax}
\title{adjust the obj object parameters for the "sw-min-max" normalization method}
\usage{
\method{fit}{ts_swminmax}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
The updated obj object
}
\description{
It takes as parameters obj object and the data
}
\examples{
trans <- dal_transform()
}
