% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_zscore.R
\name{fit.zscore}
\alias{fit.zscore}
\title{fit z-score normalization}
\usage{
\method{fit}{zscore}(obj, data, ...)
}
\arguments{
\item{obj}{z-score transformation object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
fitted z-score transformation object
}
\description{
Compute z-score normalization parameters
}
\examples{
# see ?zscore
}
