% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_fit.R
\name{fit_curvature}
\alias{fit_curvature}
\title{Fit}
\usage{
fit_curvature()
}
\value{
obj
}
\description{
In R, the fit function is a generic function used to fit a model to a data set. It is a fundamental function in many machine learning algorithms and statistical models. The fit function takes a data set and a model specification as input and returns an object containing the fitted model parameters and other relevant information
}
\examples{
trans <- dal_transform()
}
