% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{inverse_transform.default}
\alias{inverse_transform.default}
\title{Default inverse transform}
\usage{
\method{inverse_transform}{default}(obj, ...)
}
\arguments{
\item{obj}{a dal_transform object.}

\item{...}{optional arguments.}
}
\value{
It simply returns the input data which indicates that no transformation
}
\description{
Reverses the transformation applied to data.
}
