% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{inverse_transform.default}
\alias{inverse_transform.default}
\title{dal_base object}
\usage{
\method{inverse_transform}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
Simply returns NULL
}
\description{
It receives as parameter the object obj, ...
}
\examples{
trans <- dal_transform()
}
