% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{inverse_transform.minmax}
\alias{inverse_transform.minmax}
\title{Inverse Transform minmax}
\usage{
\method{inverse_transform}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{minmax object}

\item{data}{transformed dataset}

\item{...}{optional arguments}
}
\value{
inversed transformed dataset
}
\description{
The inverse transform of normalized data to original space
}
\examples{
trans <- dal_transform()
}
