% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_transform.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Inverse Transform}
\usage{
inverse_transform(obj, ...)
}
\arguments{
\item{obj}{a dal_transform object.}

\item{...}{optional arguments.}
}
\value{
dataset inverse transformed.
}
\description{
Reverses the transformation applied to data.
}
\examples{
#See ?minmax for an example of transformation
}
