% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Inverse Transform}
\usage{
inverse_transform(obj, ...)
}
\arguments{
\item{obj}{object: The transformed time series dataset.}

\item{...}{optional arguments.}
}
\value{
The time series dataset in its original scale.
}
\description{
Reverses the transformation applied to a time series dataset using the transform() function.
}
\examples{
trans <- dal_transform()
}
