% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_an.R
\name{inverse_transform.ts_an}
\alias{inverse_transform.ts_an}
\title{Perform the inversion of the transformation performed by the transform.t_an function}
\usage{
\method{inverse_transform}{ts_an}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
The final result of the inverse transformation
}
\description{
It takes 3 arguments: obj, data and x
}
\examples{
trans <- dal_transform()
}
