% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_gminmax.R
\name{inverse_transform.ts_gminmax}
\alias{inverse_transform.ts_gminmax}
\title{Inverse transformation using global min-max}
\usage{
\method{inverse_transform}{ts_gminmax}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
Inverse transformed data
}
\description{
Inverse transformation using global min-max
}
\examples{
trans <- dal_transform()
}
