% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_swminmax.R
\name{inverse_transform.ts_swminmax}
\alias{inverse_transform.ts_swminmax}
\title{Do the inverse transformation of normalized data}
\usage{
\method{inverse_transform}{ts_swminmax}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
denormalized dataset
}
\description{
It takes as parameters the variables obj, data, x
}
\examples{
trans <- dal_transform()
}
