% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_an.R
\name{ma.ts_an}
\alias{ma.ts_an}
\title{Removes the first few columns from a data array}
\usage{
ma.ts_an(obj, data, func)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{func}{function to aggregate}
}
\value{
normalized sliding windows
}
\description{
It takes as parameters the variables obj, data, func
}
\examples{
trans <- dal_transform()
}
