% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_norm_minmax.R
\name{minmax}
\alias{minmax}
\title{Min-max normalization}
\usage{
minmax(minmax_list = NULL)
}
\arguments{
\item{minmax_list}{list(feat=list(min=x, max=y)...) List of lists with min and max for each feature.

\eqn{minmax = (x-min(x))/(max(x)-min(x))}}
}
\value{
returns an object of class \code{minmax}
}
\description{
The minmax performs scales data between [0,1].
}
\examples{
data(iris)
head(iris)

trans <- minmax()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
head(tiris)

itiris <- inverse_transform(trans, tiris)
head(itiris)
}
