% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_norm_minmax.R
\name{minmax}
\alias{minmax}
\title{Min-max normalization}
\usage{
minmax()
}
\value{
returns an object of class \code{minmax}
}
\description{
The minmax performs scales data between [0,1].

\eqn{minmax = (x-min(x))/(max(x)-min(x))}
}
\examples{
data(iris)
head(iris)

trans <- minmax()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
head(tiris)

itiris <- inverse_transform(trans, tiris)
head(itiris)
}
