% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{minmax}
\alias{minmax}
\title{Min-Max Normalization}
\usage{
minmax()
}
\value{
minmax object
}
\description{
The minmax() function performs Min-Max normalization on the data
}
\examples{
trans <- dal_transform()
}
