% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_minmax.R
\name{minmax}
\alias{minmax}
\title{min-max normalization}
\usage{
minmax()
}
\value{
obj
}
\description{
The minmax performs scales data between [0,1].
}
\examples{
data(iris)
trans <- minmax()
trans <- fit(trans, iris)
tiris <- transform(trans, iris)
}
