% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize()
}
\value{
obj normalization
}
\description{
The normalize() function in the DAL package performs a normalization on the input data so that each feature (column) has zero mean and unit variance. Normalizing the data is a common preprocessing step in machine learning, particularly when using algorithms that are sensitive to the scale of the input features. The resulting normalized object is returned with the "normalize" class added to it.
}
\examples{
trans <- dal_transform()
}
