% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_outliers.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(alpha = 1.5)
}
\arguments{
\item{alpha}{boxplot outlier threshold}
}
\value{
An outlier object
}
\description{
Basic class for outlier removal
}
\examples{
trans <- dal_transform()
}
