% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_outliers_gaussian.R
\name{outliers_gaussian}
\alias{outliers_gaussian}
\title{outliers_gaussian}
\usage{
outliers_gaussian(alpha = 3)
}
\arguments{
\item{alpha}{gaussian threshold (default 3)}
}
\value{
returns an outlier object
}
\description{
The outliers_gaussian class uses box-plot definition for outliers_gaussian.
An outlier is a value that is below than \eqn{\overline{x} - 3 \sigma_x} or higher than \eqn{\overline{x} + 3 \sigma_x}.
The class remove outliers_gaussian for numeric attributes.
}
\examples{
# code for outlier removal
out_obj <- outliers_gaussian() # class for outlier analysis
out_obj <- fit(out_obj, iris) # computing boundaries
iris.clean <- transform(out_obj, iris) # returning cleaned dataset

#inspection of cleaned dataset
nrow(iris.clean)

idx <- attr(iris.clean, "idx")
table(idx)
iris.outliers_gaussian <- iris[idx,]
iris.outliers_gaussian
}
