% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_cspade.R
\name{pat_cspade}
\alias{pat_cspade}
\title{cSPADE sequences}
\usage{
pat_cspade(
  support = 0.4,
  maxsize = NULL,
  maxlen = NULL,
  mingap = NULL,
  maxgap = NULL,
  quality_filter = NULL,
  control = list(verbose = TRUE)
)
}
\arguments{
\item{support}{minimum support threshold}

\item{maxsize}{maximum number of items per event}

\item{maxlen}{maximum number of events per sequence}

\item{mingap}{minimum gap between successive events}

\item{maxgap}{maximum gap between successive events}

\item{quality_filter}{optional quality filter created with \code{patutils()}}

\item{control}{list of control parameters}
}
\value{
returns a \code{pat_cspade} object
}
\description{
Sequential pattern mining using \code{arulesSequences::cspade}.
}
\examples{
if (requireNamespace("arulesSequences", quietly = TRUE)) {
 x <- arulesSequences::read_baskets(
   con = system.file("misc", "zaki.txt", package = "arulesSequences"),
   info = c("sequenceID", "eventID", "SIZE")
 )
 utils <- patutils()
 pm <- pat_cspade(
   support = 0.4,
   maxlen = 3,
   quality_filter = utils$quality_min(support = 0.5)
 )
 pm <- fit(pm, x)
 seqs <- discover(pm, x)
 eval <- evaluate(pm, seqs)
 eval$metrics
}
}
