% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_eclat.R
\name{pat_eclat}
\alias{pat_eclat}
\title{ECLAT itemsets}
\usage{
pat_eclat(
  supp = 0.5,
  minlen = 1,
  maxlen = 3,
  include = NULL,
  exclude = NULL,
  quality_filter = NULL,
  control = NULL
)
}
\arguments{
\item{supp}{minimum support threshold}

\item{minlen}{minimum itemset length}

\item{maxlen}{maximum itemset length}

\item{include}{optional vector of items allowed in the discovered itemsets}

\item{exclude}{optional vector of items forbidden in the discovered itemsets}

\item{quality_filter}{optional quality filter created with \code{patutils()}}

\item{control}{list of control parameters}
}
\value{
returns a \code{pat_eclat} object
}
\description{
Frequent itemsets using \code{arules::eclat}.
}
\examples{
if (requireNamespace("arules", quietly = TRUE)) {
 data("AdultUCI", package = "arules")
 trans <- suppressWarnings(methods::as(as.data.frame(AdultUCI), "transactions"))
 utils <- patutils()
 pm <- pat_eclat(
   supp = 0.2,
   maxlen = 3,
   include = c("sex=Male", "income=small", "marital-status=Married-civ-spouse", "race=White"),
   exclude = c("income=small"),
   quality_filter = utils$quality_min(support = 0.4),
   control = list(verbose = FALSE)
 )
 pm <- fit(pm, trans)
 itemsets <- discover(pm, trans)
 eval <- evaluate(pm, itemsets)
 eval$metrics
}
}
