% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_pattern_miner.R
\name{pattern_miner}
\alias{pattern_miner}
\title{Pattern miner}
\usage{
pattern_miner()
}
\value{
returns a \code{pattern_miner} object
}
\description{
Base class for frequent pattern and sequence mining.
}
