% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.boxplot.class}
\alias{plot.boxplot.class}
\title{plot scatter}
\usage{
plot.boxplot.class(
  data,
  class_label,
  label_x = "",
  label_y = "",
  colors = NULL
)
}
\value{
plot
}
\description{
plot scatter
}
