% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot bar graph}
\usage{
plot_bar(data, label_x = "", label_y = "", colors = NULL, alpha = 1)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{alpha}{level of transparency}
}
\value{
returns a ggplot2::ggplot graphic
}
\description{
this function displays a bar graph from a data frame containing x-axis categories using ggplot2.
}
\examples{
#summarizing iris dataset
data <- iris |> dplyr::group_by(Species) |>
dplyr::summarize(Sepal.Length=mean(Sepal.Length))
head(data)

#ploting data
grf <- plot_bar(data, colors="blue")
plot(grf)
}
