% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot boxplot}
\usage{
plot_boxplot(data, label_x = "", label_y = "", colors = NULL, barwidth = 0.25)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{barwidth}{width of bar}
}
\value{
returns a ggplot2::ggplot graphic
}
\description{
this function displays a boxplot graph from a data frame containing x-axis categories and numeric values using ggplot2.
}
\examples{
grf <- plot_boxplot(iris, colors="white")
plot(grf)
}
