% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_boxplot_class}
\alias{plot_boxplot_class}
\title{Boxplot per class}
\usage{
plot_boxplot_class(
  data,
  class_label,
  label_x = "",
  label_y = "",
  colors = NULL
)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{class_label}{name of attribute for class label}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
returns a ggplot graphic
}
\description{
This function generates boxplots grouped by a specified class label from a data frame containing numeric values using ggplot2.
}
\examples{
grf <- plot_boxplot_class(iris |> dplyr::select(Sepal.Width, Species),
 class = "Species", colors=c("red", "green", "blue"))
plot(grf)
}
