% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_density}
\alias{plot_density}
\title{Plot density}
\usage{
plot_density(
  data,
  label_x = "",
  label_y = "",
  colors = NULL,
  bin = NULL,
  alpha = 0.25
)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{bin}{bin width for density estimation}

\item{alpha}{level of transparency}
}
\value{
returns a ggplot graphic
}
\description{
This function generates a density plot from a data frame containing numeric values using ggplot2.
If the data frame has multiple columns, densities can be grouped and plotted.
}
\examples{
grf <- plot_density(iris |> dplyr::select(Sepal.Width), colors="blue")
plot(grf)
}
