% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_density}
\alias{plot_density}
\title{plot density}
\usage{
plot_density(
  data,
  label_x = "",
  label_y = "",
  colors = NULL,
  bin = NULL,
  alpha = 0.25
)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{bin}{bin width}

\item{alpha}{level of transparency}
}
\description{
plot density
}
\examples{
trans <- dal_transform()
}
