% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{plot lollipop}
\usage{
plot_lollipop(
  data,
  colors,
  xlabel = "",
  ylabel = "",
  size_text = 3,
  size_ball = 8,
  alpha_ball = 0.2,
  min_value = 0,
  max_value_gap = 1,
  flip = TRUE
)
}
\value{
plot
}
\description{
plot lollipop
}
