% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pieplot}
\alias{plot_pieplot}
\title{Plot pie}
\usage{
plot_pieplot(
  data,
  label_x = "",
  label_y = "",
  colors = NULL,
  textcolor = "white",
  bordercolor = "black"
)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}

\item{textcolor}{text color}

\item{bordercolor}{border color}
}
\value{
returns a ggplot graphic
}
\description{
This function creates a pie chart using ggplot2.
}
\examples{
#summarizing iris dataset
data <- iris |> dplyr::group_by(Species) |>
 dplyr::summarize(Sepal.Length=mean(Sepal.Length))
head(data)

#ploting data
grf <- plot_pieplot(data, colors=c("red", "green", "blue"))
plot(grf)
}
