% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_radar}
\alias{plot_radar}
\title{plot radar}
\usage{
plot_radar(data, label_x = "", label_y = "", colors = NULL)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
ggplot graphic
}
\description{
plot radar
}
\examples{
trans <- dal_transform()
}
