% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot scatter}
\usage{
plot_scatter(
  series,
  label_series = "",
  label_x = "",
  label_y = "",
  colors = NULL
)
}
\value{
plot
}
\description{
plot scatter
}
