% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_scatter}
\alias{plot_scatter}
\title{scatter graph}
\usage{
plot_scatter(
  series,
  label_series = "",
  label_x = "",
  label_y = "",
  colors = NULL
)
}
\arguments{
\item{series}{data.frame contain x, value, and variable}

\item{label_series}{title label}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
ggplot graphic
}
\description{
scatter graph
}
\examples{
trans <- dal_transform()
}
