% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_scatter}
\alias{plot_scatter}
\title{scatter graph}
\usage{
plot_scatter(data, label_x = "", label_y = "", colors = NULL)
}
\arguments{
\item{data}{data.frame contain x, value, and variable}

\item{label_x}{x-axis label}

\item{label_y}{y-axis label}

\item{colors}{color vector}
}
\value{
ggplot graphic
}
\description{
scatter graph
}
\examples{
grf <- plot_scatter(iris |> dplyr::select(x = Sepal.Length,
 value = Sepal.Width, variable = Species),
 label_x = "Sepal.Length", label_y = "Sepal.Width",
 colors=c("red", "green", "blue"))
 plot(grf)
}
