% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{predict.tsreg_sw}
\alias{predict.tsreg_sw}
\title{Performs prediction of values for a time series}
\usage{
\method{predict}{tsreg_sw}(obj, x, steps_ahead = 1)
}
\arguments{
\item{steps_ahead}{}
}
\value{
A vector with forecasts for each time period
}
\description{
The function receives three variables as a parameter, which are obj, x and steps_ahead
}
\details{
If the steps_ahead parameter is equal to 1, the function takes a set of data x and returns a forecast y for the next time period. If steps_ahead is greater than 1, the function iterates over the number of time periods to be predicted and performs the same process for each one
}
