% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_predictor.R
\name{predictor}
\alias{predictor}
\title{Predictor (base for classification/regression)}
\usage{
predictor()
}
\value{
returns a predictor object
}
\description{
Ancestor class for supervised predictors (classification and regression).
Provides a default \code{fit()} to record feature names and proxies \code{action()} to \code{predict()}.

An example predictor is a decision tree classifier (\code{cla_dtree}).
}
\examples{
#See ?cla_dtree for a classification example using a decision tree
}
