% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_svm.R
\name{reg_svm}
\alias{reg_svm}
\title{Support Vector Machine (SVM) Regression}
\usage{
reg_svm(
  attribute,
  epsilon = seq(0, 1, 0.2),
  cost = seq(20, 100, 20),
  kernel = "radial"
)
}
\description{
Support Vector Machine (SVM) Regression
}
