% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{register_log.default}
\alias{register_log.default}
\title{dal_base object}
\usage{
\method{register_log}{default}(obj, msg = "", ref = "")
}
\arguments{
\item{obj}{object}

\item{msg}{message to debug}

\item{ref}{reference}
}
\value{
The obj object updated with log record information
}
\description{
It takes as parameters the obj object and the variable msg. This function logs a log message to an object and returns the updated object itself
}
\examples{
trans <- dal_transform()
}
