% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{register_log}
\alias{register_log}
\title{Register log details}
\usage{
register_log(obj, msg, ref)
}
\arguments{
\item{obj}{object}

\item{msg}{string: a message to the log.}

\item{ref}{.}
}
\value{
obj
}
\description{
It takes as parameters the variables obj, msg and ref. It serves to direct the appropriate method call, depending on the type of object passed as the obj argument.
}
\examples{
trans <- dal_transform()
}
