% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_regression.R
\name{regression}
\alias{regression}
\title{Regression base class}
\usage{
regression(attribute)
}
\arguments{
\item{attribute}{attribute target to model building}
}
\value{
returns a regression object
}
\description{
Ancestor class for regression models. Stores the target attribute and provides common evaluation metrics.
}
\examples{
#See ?reg_dtree for a regression example using a decision tree
}
