% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_regression.R
\name{regression}
\alias{regression}
\title{Regression}
\usage{
regression(attribute)
}
\arguments{
\item{attribute}{attribute target to model building}
}
\value{
returns a regression object
}
\description{
Ancestor class for regression problems.
This ancestor class is used to define and manage the target attribute for regression tasks.
}
\examples{
#See ?reg_dtree for a regression example using a decision tree
}
