% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_regression.R
\name{regression}
\alias{regression}
\title{Regression base class}
\usage{
regression(attribute)
}
\arguments{
\item{attribute}{attribute target to model building}
}
\value{
obj
}
\description{
Regression base class
}
\examples{
trans <- dal_transform()
}
