% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_sample_random.R
\name{sample_random}
\alias{sample_random}
\title{Sample Random}
\usage{
sample_random()
}
\value{
returns an object of class `sample_random
}
\description{
The sample_random function in R is used to
generate a random sample of specified size from a given data set.
}
\examples{
#using random sampling
sample <- sample_random()
tt <- train_test(sample, iris)

# distribution of train
table(tt$train$Species)

# preparing dataset into four folds
folds <- k_fold(sample, iris, 4)

# distribution of folds
tbl <- NULL
for (f in folds) {
 tbl <- rbind(tbl, table(f$Species))
}
head(tbl)
}
