% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample.R
\name{sample_random}
\alias{sample_random}
\title{Sample Random}
\usage{
sample_random()
}
\value{
obj
}
\description{
The sample_random function in R is used to generate a random sample of specified size from a given data set.
}
\examples{
trans <- dal_transform()
}
