% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample_strat.R
\name{sample_stratified}
\alias{sample_stratified}
\title{sample_stratified}
\usage{
sample_stratified(attribute)
}
\description{
The sample_stratified function in R is used to generate a stratified random sample from a given dataset. Stratified sampling is a statistical method that is used when the population is divided into non-overlapping subgroups or strata, and a sample is selected from each stratum to represent the entire population. In stratified sampling, the sample is selected in such a way that it is representative of the entire population and the variability within each stratum is minimized.
}
\details{
The sample_stratified function requires the following inputs: data: A data frame containing the data to be sampled.; strata: A character vector containing the name of the variable(s) to be used as strata.; size: An integer value specifying the size of the sample to be drawn.; replace: A logical value indicating whether sampling should be done with replacement or without replacement.; seed: An optional integer value used to set the seed for the random number generator.
}
