% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_tune.R
\name{select_hyper.cla_tune}
\alias{select_hyper.cla_tune}
\title{selection of hyperparameters}
\usage{
\method{select_hyper}{cla_tune}(obj, hyperparameters)
}
\arguments{
\item{obj}{an object representing the model or tuning process}

\item{hyperparameters}{a dataframe with columns \code{key} (hyperparameter configuration) and \code{metric} (classification metric)}
}
\value{
returns a optimized key number of hyperparameters
}
\description{
Selects the optimal hyperparameter by maximizing the average classification metric. It wraps dplyr library.
}
