% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_tune.R
\name{select_hyper}
\alias{select_hyper}
\title{Selection hyper parameters}
\usage{
select_hyper(obj, hyperparameters)
}
\arguments{
\item{obj}{the object or model used for hyperparameter selection.}

\item{hyperparameters}{data set with hyper parameters and quality measure from execution}
}
\value{
returns the index of selected hyper parameter
}
\description{
Selects the optimal hyperparameters from a dataset resulting from k-fold cross-validation
}
