% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_tune.R
\name{select_hyper}
\alias{select_hyper}
\title{select hyper parameters}
\usage{
select_hyper(obj, hyperparameters)
}
\arguments{
\item{obj}{object}

\item{hyperparameters}{data set with hyper parameters.}
}
\value{
index of hyperparameter
}
\description{
select hyper parameters
}
\examples{
trans <- dal_transform()
}
