% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_kmeans.R
\name{set_params.cluster_kmeans}
\alias{set_params.cluster_kmeans}
\title{Updates the parameters of cluster_kmeans}
\usage{
\method{set_params}{cluster_kmeans}(obj, params)
}
\value{
The obj template object updated with the new parameters
}
\description{
It takes as input the obj model object and a set of params parameters to update.
}
\details{
The function checks that each parameter specified in params is not null and, if not null, updates the corresponding parameter in the object obj
}
