% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_pam.R
\name{set_params.cluster_pam}
\alias{set_params.cluster_pam}
\title{Updates the parameters of cluster_pam}
\usage{
\method{set_params}{cluster_pam}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
The obj template object updated with the new parameters
}
\description{
It takes as input the obj model object and a set of params parameters to update.
}
