% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{set_params.dal_base}
\alias{set_params.dal_base}
\title{Set parameters values for dal_base}
\usage{
\method{set_params}{dal_base}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
The dal_base object updated with the new parameter values
}
\description{
It receives as input a dal_base object (obj) and a set of parameters (params)
}
