% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{set_params.default}
\alias{set_params.default}
\title{Default Assign parameters}
\usage{
\method{set_params}{default}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
returns the object unchanged
}
\description{
Default method for \code{set_params} (returns object unchanged).
}
