% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{set_params.default}
\alias{set_params.default}
\title{Defines a default method for the set_params function}
\usage{
\method{set_params}{default}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
The obj object
}
\description{
This function receives the obj and params variables as parameters.
}
