% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_base.R
\name{set_params.default}
\alias{set_params.default}
\title{Assign parameters}
\usage{
\method{set_params}{default}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
obj
}
\description{
This function receives the obj and params variables as parameters.
It returns the obj as it is.
}
