% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{set_params}
\alias{set_params}
\title{Update an object's parameters}
\usage{
set_params(obj, params)
}
\arguments{
\item{obj}{object: .}
}
\description{
The function receives the obj and params variables as parameters
}
\details{
This function serves as a "skeleton" for implementing different set_params methods for different object classes
}
