% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_elm.R
\name{set_params.ts_elm}
\alias{set_params.ts_elm}
\title{Set specific parameters of a ts_elm object}
\usage{
\method{set_params}{ts_elm}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
The object obj updated with the new parameters
}
\description{
This function takes two arguments: obj, which is an object of the ts_elm class, and params, which is a list of parameters to be updated
}
\examples{
trans <- dal_transform()
}
