% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mlp.R
\name{set_params.ts_mlp}
\alias{set_params.ts_mlp}
\title{Update ts_mlp object parameters}
\usage{
\method{set_params}{ts_mlp}(obj, params)
}
\arguments{
\item{obj}{object}

\item{params}{parameters}
}
\value{
ts_mlp object updated with new parameter values
}
\description{
It takes two arguments: the ts_mlp object to be updated (obj) and a list object containing the new parameter values to be updated (params)
}
