% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_rf.R
\name{set_params.ts_rf}
\alias{set_params.ts_rf}
\title{Allows parameters of a ts_rf object to be updated with new values}
\usage{
\method{set_params}{ts_rf}(obj, params)
}
\arguments{
\item{params}{}
}
\value{
The ts_rf object updated with the new parameter values
}
\description{
It receives as input a ts_rf object (obj) and a set of parameters (params)
}
\details{
If the parameter set contains an entry for nodesize, the corresponding value is assigned to the ts_rf object. Likewise, if the parameter set contains an entry for ntree, the corresponding value is assigned to the ts_rf object
}
