% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{sMAPE.ts}
\alias{sMAPE.ts}
\title{sMAPE}
\usage{
sMAPE.ts(actual, prediction)
}
\arguments{
\item{actual}{real observations}

\item{prediction}{predicted observations}
}
\value{
returns the sMAPE between the actual and prediction vectors
}
\description{
Compute the symmetric mean absolute percent error (sMAPE)
}
