% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smoothing}
\usage{
smoothing(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
obj
}
\description{
Smoothing is a statistical technique that is used to reduce the noise in a signal or a dataset by removing the high-frequency components. It involves applying a smoothing function or algorithm to the dataset that results in a smoother curve that is easier to analyze. The smoothing technique can be applied to any type of data, including time-series data, images, and sound signals.
}
\examples{
trans <- dal_transform()
}
