% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smoothing}
\usage{
smoothing(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
obj
}
\description{
Smoothing is a statistical technique used to reduce
the noise in a signal or a dataset by removing the high-frequency components.
The smoothing level is associated with the number of bins used.
There are alternative methods to establish the smoothing:
equal interval, equal frequency, and clustering.
}
\examples{
trans <- dal_transform()
}
