% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_cluster}
\alias{smoothing_cluster}
\title{Smoothing by cluster}
\usage{
smoothing_cluster(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
obj
}
\description{
The function smoothing_cluster() is used to perform smoothing of data by cluster. This function takes as input a numeric vector, which is divided into clusters using the k-means algorithm. The mean of each cluster is then calculated and used as the smoothed value for all observations within that cluster.
}
\examples{
trans <- dal_transform()
}
