% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_evaluation}
\alias{smoothing_evaluation}
\title{Smoothing by evaluation}
\usage{
smoothing_evaluation(data, attribute)
}
\arguments{
\item{data}{dataset}

\item{attribute}{attribute target to model building}
}
\value{
data entropy
}
\description{
The smoothing_evaluation() function performs a smoothing evaluation using cross-validation. It calculates the mean squared error (MSE) for each smoothing parameter and returns a data frame with the results. The function uses the smoothing technique specified by the user to smooth the data.
}
\examples{
trans <- dal_transform()
}
