% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_smoothing_freq.R
\name{smoothing_freq}
\alias{smoothing_freq}
\title{Smoothing by Freq}
\usage{
smoothing_freq(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
returns an object of class \code{smoothing_freq}
}
\description{
The 'smoothing_freq' function is used to smooth a given time series data by aggregating observations within a fixed frequency.
}
\examples{
data(iris)
obj <- smoothing_freq(n = 2)
obj <- fit(obj, iris$Sepal.Length)
sl.bi <- transform(obj, iris$Sepal.Length)
table(sl.bi)
obj$interval

entro <- evaluate(obj, as.factor(names(sl.bi)), iris$Species)
entro$entropy
}
