% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing_freq.R
\name{smoothing_freq}
\alias{smoothing_freq}
\title{Smoothing by Freq}
\usage{
smoothing_freq(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
obj
}
\description{
The 'smoothing_freq' function is used to smooth a given time series data by aggregating observations within a fixed frequency.
}
\examples{
trans <- dal_transform()
}
