% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_inter}
\alias{smoothing_inter}
\title{Smoothing by interval}
\usage{
smoothing_inter(n)
}
\arguments{
\item{n}{number of bins}
}
\value{
obj
}
\description{
The "smoothing by interval" function is used to apply a smoothing technique to a vector or time series data using a moving window approach.
}
\examples{
trans <- dal_transform()
}
