% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_smoothing.R
\name{smoothing_inter}
\alias{smoothing_inter}
\title{Smoothing by interval}
\usage{
smoothing_inter(n)
}
\description{
The "smoothing by interval" function is used to apply a smoothing technique to a vector or time series data using a moving window approach.
}
\details{
The function takes in three arguments:
data: a numeric vector or time series data to be smoothed.
window_size: an integer value specifying the size of the moving window. It should be an odd number to ensure a centered window.
method: a character string specifying the smoothing method to be used. It can be one of "simple", "linear", "exponential", "spline", "kernel", or "loess".
}
