% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{start_log.default}
\alias{start_log.default}
\title{dal_base object}
\usage{
\method{start_log}{default}(obj)
}
\arguments{
\item{obj}{object: .}
}
\value{
The object with a new attribute "log_time",
}
\description{
It takes as parameter the obj object
}
\details{
This function is a method to start logging on an object. The default method assigns the current time to the object and returns it
}
