% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{start_log}
\alias{start_log}
\title{Start logging on an obj object}
\usage{
start_log(obj)
}
\arguments{
\item{obj}{object}
}
\value{
obj
}
\description{
It takes as parameter an obj object
}
\examples{
trans <- dal_transform()
}
