% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{[.ts_data}
\alias{[.ts_data}
\title{Extract a subset of a time series stored in an object}
\usage{
\method{[}{ts_data}(x, i, j, ...)
}
\arguments{
\item{x}{input variable}

\item{i}{row i}

\item{j}{column j}

\item{...}{optional arguments}
}
\value{
A new ts_data object
}
\description{
Receives as parameters the variables x, i, j ...
}
\examples{
trans <- dal_transform()
}
