% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{[.ts_data}
\alias{[.ts_data}
\title{Extract a subset of a time series stored in an object}
\usage{
\method{[}{ts_data}(x, i, j, ...)
}
\value{
A new ts_data object
}
\description{
Receives as parameters the variables x, i, j ...
}
\details{
The i and j arguments specify the rows and columns to be selected, respectively. The argument ... lets you pass other arguments to the unclass function
}
