% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample.R
\name{train_test}
\alias{train_test}
\title{training and test partition object}
\usage{
train_test(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments.}
}
\value{
train and test sets
}
\description{
training and test partition object
}
\examples{
trans <- dal_transform()
}
