% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_sample_random.R
\name{train_test_from_folds}
\alias{train_test_from_folds}
\title{k-fold training and test partition object}
\usage{
train_test_from_folds(folds, k)
}
\arguments{
\item{folds}{data partitioned into folds}

\item{k}{k-fold for test set, all reminder for training set}
}
\value{
train and test folds
}
\description{
k-fold training and test partition object
}
