% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{transform.default}
\alias{transform.default}
\title{Transform}
\usage{
\method{transform}{default}(obj, ...)
}
\arguments{
\item{obj}{a dal_transform object.}

\item{...}{optional arguments}
}
\value{
It simply returns NULL, which indicates that no transformation
}
\description{
Executes the default transformation of an object.
}
