% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{transform.default}
\alias{transform.default}
\title{dal_base object}
\usage{
\method{transform}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional arguments}
}
\value{
It simply returns NULL, which indicates that no transforms are applied
}
\description{
A default function that defines the default behavior of the transform function for objects of class dal_transform
}
\examples{
trans <- dal_base()
}
