% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{transform.minmax}
\alias{transform.minmax}
\title{Transform MinMax}
\usage{
\method{transform}{minmax}(obj, data)
}
\description{
The transform() function in the DAL package transforms the input data based on the normalization set stored in the object generated by the fit.minmax() function. The function applies the min-max scaling method on the input data to bring the values of each feature (column) within the range of 0 to 1, based on the minimum and maximum values of the feature in the normalization set. If the minimum and maximum values of the feature are the same, the function sets the values in that feature to 0.
}
\details{
The transform() function takes the normalization set stored in the input object and applies it to the input data. The normalization set contains the minimum and maximum values of each feature in the original data set. The function applies the min-max scaling method to each feature in the input data by subtracting the minimum value of the feature from all values in that feature and dividing the result by the difference between the maximum and minimum values of that feature in the normalization set. If the minimum and maximum values of the feature are the same, the function sets the values in that feature to 0.
}
