% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_norm_minmax.R
\name{transform.minmax}
\alias{transform.minmax}
\title{transform min-max normalization}
\usage{
\method{transform}{minmax}(obj, data, ...)
}
\arguments{
\item{obj}{min-max transformation object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
transformed dataset
}
\description{
Scale data using minmax normalization
}
\examples{
# see ?minmax
}
