% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_outliers.R
\name{transform.outliers}
\alias{transform.outliers}
\title{Remove outliers from data}
\usage{
\method{transform}{outliers}(obj, data, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
returns the output dataset after applying the transformation and adds an "idx" attribute that contains a boolean vector indicating which rows were removed from the original dataset
}
\description{
Remove outliers from data
}
\examples{
trans <- dal_transform()
}
