% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_transform.R
\name{transform}
\alias{transform}
\title{Transform}
\usage{
transform(obj, ...)
}
\arguments{
\item{obj}{a \code{dal_transform} object.}

\item{...}{optional arguments.}
}
\value{
returns a transformed data.
}
\description{
Generic to apply a transformation to data.
}
\examples{
#See ?minmax for an example of transformation
}
