% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_transform.R
\name{transform}
\alias{transform}
\title{Transform}
\usage{
transform(obj, ...)
}
\arguments{
\item{obj}{object: a dal_transform object to apply the transformation to.}

\item{...}{optional arguments.}
}
\value{
the transformed time series data.
}
\description{
Defines the kind of transformation to be set over a time series.
}
\examples{
trans <- dal_transform()
}
