% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_an.R
\name{transform.ts_an}
\alias{transform.ts_an}
\title{Normalize the input data}
\usage{
\method{transform}{ts_an}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
Normalized data along with annual moving average value
}
\description{
It takes 3 arguments: obj, data and x
}
\examples{
trans <- dal_transform()
}
