% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_diff.R
\name{transform.ts_diff}
\alias{transform.ts_diff}
\title{Transform the data}
\usage{
\method{transform}{ts_diff}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
Transformed full data matrix
}
\description{
It takes as parameters the variables obj, data and x
}
\examples{
trans <- dal_transform()
}
