% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_gminmax.R
\name{transform.ts_gminmax}
\alias{transform.ts_gminmax}
\title{Normalize time series data}
\usage{
\method{transform}{ts_gminmax}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
normalized data
}
\description{
It takes as parameters the variables obj, data and x
}
\examples{
trans <- dal_transform()
}
