% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_swminmax.R
\name{transform.ts_swminmax}
\alias{transform.ts_swminmax}
\title{Normalizes each line of the input data to the range [0,1]}
\usage{
\method{transform}{ts_swminmax}(obj, data, x = NULL, ...)
}
\arguments{
\item{obj}{object}

\item{data}{dataset}

\item{x}{input variable}

\item{...}{optional arguments}
}
\value{
transformed dataset
}
\description{
It takes as parameters the variables obj, data, x
}
\examples{
trans <- dal_transform()
}
