% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{transform.ts_transform}
\alias{transform.ts_transform}
\title{Transform data to objects of class st_transform}
\usage{
\method{transform}{ts_transform}(obj, data)
}
\value{
The data given as an argument
}
\description{
The function receives as arguments the variables obj and data
}
