% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{transform.zscore}
\alias{transform.zscore}
\title{Transform z-score}
\usage{
\method{transform}{zscore}(obj, data, ...)
}
\arguments{
\item{obj}{z-score object}

\item{data}{dataset}

\item{...}{optional arguments}
}
\value{
transformed dataset
}
\description{
Scale data using miz-score normalization
}
