% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_conv1d.R
\name{ts_conv1d}
\alias{ts_conv1d}
\title{Conv1D}
\usage{
ts_conv1d(preprocess = NA, input_size = NA, epochs = 10000L)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}

\item{epochs}{maximum number of epochs}
}
\value{
returns a \code{ts_conv1d} object.
}
\description{
Creates a time series prediction object that uses the Conv1D.
It wraps the pytorch library.
}
\examples{
#See an example of using `ts_conv1d` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/timeseries/ts_conv1d.md
}
