% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_data}
\alias{ts_data}
\title{ts_data}
\usage{
ts_data(y, sw = 1)
}
\arguments{
\item{y}{output variable}

\item{sw}{integer: sliding window size.}
}
\value{
returns a \code{ts_data} object.
}
\description{
Time series data structure used in DAL Toolbox.
It receives a vector (representing a time series) or
a matrix \code{y} (representing a sliding windows).
Internal ts_data is matrix of sliding windows with size \code{sw}.
If sw equals to zero, it store a time series as a single matrix column.
}
\examples{
data(sin_data)
head(sin_data)

data <- ts_data(sin_data$y)
ts_head(data)

data10 <- ts_data(sin_data$y, 10)
ts_head(data10)
}
