% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_norm_diff.R
\name{ts_diff}
\alias{ts_diff}
\title{Time Series Diff}
\usage{
ts_diff(remove_outliers = TRUE)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_diff} object.
}
\description{
It receives as parameter the variable remove_outliters. This function calculates the difference between the values of a time series
}
\examples{
trans <- dal_transform()
}
