% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ts_ean}
\alias{ts_ean}
\title{Time Series Adaptive Normalization (Exponential Moving Average - EMA)}
\usage{
ts_ean(remove_outliers = TRUE, nw = 0)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_ean} object.
}
\description{
It takes 2 parameters: remove_outliers and nw
}
\details{
The function uses the "emean" function to calculate the exponentially weighted average. The time series can be pre-processed to remove outliers (remove_outliers) and can also have a specified window size for smoothing (nw)
}
