% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_lstm.R
\name{ts_lstm}
\alias{ts_lstm}
\title{LSTM}
\usage{
ts_lstm(preprocess = NA, input_size = NA, epochs = 10000L)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}

\item{epochs}{maximum number of epochs}
}
\value{
returns a \code{ts_lstm} object.
}
\description{
Creates a time series prediction object that uses the LSTM.
It wraps the pytorch library.
}
\examples{
#See an example of using `ts_ts_lstmconv1d` at this
#https://github.com/cefet-rj-dal/daltoolbox/blob/main/timeseries/ts_lstm.md
}
