% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_main.R
\name{ts_plot_pred}
\alias{ts_plot_pred}
\title{Plot a time series chart}
\usage{
ts_plot_pred(
  x = NULL,
  y,
  yadj,
  ypred = NULL,
  color = "black",
  label_x = "",
  label_y = ""
)
}
\arguments{
\item{x}{time index}

\item{y}{time series}

\item{yadj}{adjustment of time series}

\item{ypred}{prediction of the time series}

\item{color}{color for the time series}

\item{label_x}{x-axis title}

\item{label_y}{y-axis title}
}
\description{
The function receives six variables as a parameter, which are obj and y, yadj, main and xlabels. The graph is plotted with 3 lines: the original series (in black), the adjusted series (in blue) and the predicted series (in green)
}
