% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_projection.R
\name{ts_projection}
\alias{ts_projection}
\title{Time Series Projection}
\usage{
ts_projection(ts)
}
\arguments{
\item{ts}{matrix or data.frame containing the time series.}
}
\value{
returns a \code{ts_projection} object.
}
\description{
Separates a \code{ts_data} object into input and output components for time series analysis.
This function is useful for preparing data for modeling, where the input and output variables are extracted from a time series dataset.
}
\examples{
#setting up a ts_data
data(sin_data)
ts <- ts_data(sin_data$y, 10)

io <- ts_projection(ts)

#input data
ts_head(io$input)

#output data
ts_head(io$output)
}
