% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{ts_projection}
\alias{ts_projection}
\title{Time Series Projection}
\usage{
ts_projection(ts)
}
\arguments{
\item{ts}{matrix or data.frame containing the time series.}
}
\value{
a \code{ts_projection} object.
}
\description{
Separates the \code{ts_data} into input and output.
}
\examples{
#setting up a ts_data
data(sin_data)
ts <- ts_data(sin_data$y, 10)

io <- ts_projection(ts)

#input data
ts_head(io$input)

#output data
ts_head(io$output)
}
