% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{ts_reg}
\alias{ts_reg}
\title{TSReg}
\usage{
ts_reg()
}
\value{
returns \code{ts_reg} object
}
\description{
Time Series Regression directly from time series
Ancestral class for non-sliding windows implementation.
}
\examples{
#This is an abstract class.
}
