% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_regsw.R
\name{ts_regsw}
\alias{ts_regsw}
\title{TSRegSW}
\usage{
ts_regsw(preprocess = NA, input_size = NA)
}
\arguments{
\item{preprocess}{normalization}

\item{input_size}{input size for machine learning model}
}
\value{
returns a \code{ts_regsw} object
}
\description{
Time Series Regression from Sliding Windows.
Ancestral class for Machine Learning Implementation.
}
\examples{
#This is an abstract class.
}
