% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_rf.R
\name{ts_rf}
\alias{ts_rf}
\title{Time Series Random Forest}
\usage{
ts_rf(preprocess = NA, input_size = NA, nodesize = 5, ntree = 20)
}
\value{
a \code{ts_rf} object.
}
\description{
The function receives as arguments the variables preprocess, input_size, nodesize and ntree
}
\details{
Create an object that contains information about the preprocessing (if applicable), the input window size (input_size), the node size (nodesize), and the number of trees (ntree) for the model
}
