% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_filter.R
\name{tsfil_ma}
\alias{tsfil_ma}
\title{Time Series Moving Average}
\usage{
tsfil_ma(ma = 3)
}
\value{
a \code{tsfil_ma} object.
}
\description{
Used to smooth out fluctuations and reduce noise in a time series.
}
