% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{tshead}
\alias{tshead}
\title{Extract the first observations from a time series}
\usage{
tshead(x, n = 6L, ...)
}
\arguments{
\item{n}{integer: size of test data.}
}
\value{
The first n observations of a time series x
}
\description{
The function takes as arguments the variables x, n (default = 6L), ...
}
\details{
The function uses the head function to extract the first n observations of the x-time series, after turning it into a non-class object via the unclass function
}
