% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_normalization.R
\name{zscore}
\alias{zscore}
\title{Z-score Normalization}
\usage{
zscore(nmean = 0, nsd = 1)
}
\description{
The zscore() function in the DAL package performs a normalization on the input data so that each feature (column) has zero mean and standard deviation of nsd. Normalizing the data is a common preprocessing step in machine learning, particularly when using algorithms that are sensitive to the scale of the input features. The resulting zscore normalized object is returned with the "zscore" class added to it.
}
\details{
The function takes two optional arguments:

nmean: A numeric value that specifies the desired mean of the normalized data. The default value is 0.

nsd: A numeric value that specifies the desired standard deviation of the normalized data. The default value is 1.
}
